% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Indexes.R
\name{createCharacterizationIndexes}
\alias{createCharacterizationIndexes}
\title{A function to add indexes to the Characterization results}
\usage{
createCharacterizationIndexes(connectionHandler, schema, cTablePrefix = "c_")
}
\arguments{
\item{connectionHandler}{A connection handler that connects to the database and extracts sql queries.  Create a connection handler via `ResultModelManager::ConnectionHandler$new()`.}

\item{schema}{The result database schema (e.g., 'main' for sqlite)}

\item{cTablePrefix}{The prefix used for the characterization results tables}
}
\description{
A function to add indexes to the Characterization results
}
\details{
Specify the connectionHandler, the schema and the prefixes
}
\seealso{
Other Indexes:
\code{\link[=createCohortIndexes]{createCohortIndexes()}},
\code{\link[=createIncidenceIndexes]{createIncidenceIndexes()}},
\code{\link[=createSccsIndexes]{createSccsIndexes()}}
}
\concept{Indexes}
