% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SccPlots.R
\name{createSccSignalsTable}
\alias{createSccSignalsTable}
\title{Create a self controlled cohort signal discovery table}
\usage{
createSccSignalsTable(data)
}
\arguments{
\item{data}{A data.frame returned by \code{\link{getSccSignals}}}
}
\value{
A gt table
}
\description{
Create a formatted gt table of the signal discovery grid returned by
\code{\link{getSccSignals}}
}
\seealso{
Other SelfControlledCohort: 
\code{\link{getSccAnalysisSettings}()},
\code{\link{getSccDiagnosticsData}()},
\code{\link{getSccEstimation}()},
\code{\link{getSccMetaEstimation}()},
\code{\link{getSccNegativeControlEstimates}()},
\code{\link{getSccOutcomes}()},
\code{\link{getSccSignals}()},
\code{\link{getSccSummaryStats}()},
\code{\link{getSccTargets}()},
\code{\link{plotSccBoxPlot}()},
\code{\link{plotSccForest}()},
\code{\link{plotSccSystematicError}()}
}
\concept{SelfControlledCohort}
