% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SccQueries.R
\name{.sccQuoteLiterals}
\alias{.sccQuoteLiterals}
\title{Quote and join values for use in an SQL IN clause}
\usage{
.sccQuoteLiterals(values)
}
\arguments{
\item{values}{A vector of character values}
}
\value{
A single comma separated and quoted string
}
\description{
Quote and join values for use in an SQL IN clause
}
