% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{formatTargetPopulation}
\alias{formatTargetPopulation}
\title{output a string summarizing the target population}
\usage{
formatTargetPopulation(popSetting)
}
\arguments{
\item{popSetting}{the target population setting}
}
\value{
An object of class `knitr_kable` that will show the data via a nice html table
}
\description{
This returns a character that describes the target population
}
\details{
Input the target population setting list
}
\examples{
targetText <- formatTargetPopulation(
list(
limitToFirstInNDays = 365,
minPriorObservation = 365,
nestingName = 'Indication',
minAge = 12, maxAge = 9999,
studyEnd = 'Mar 2012', studyStart = NA,
genderConceptIds = NA
)
)

}
\seealso{
Other helper:
\code{\link[=addTarColumn]{addTarColumn()}},
\code{\link[=formatBinaryCovariateName]{formatBinaryCovariateName()}},
\code{\link[=getExampleConnectionDetails]{getExampleConnectionDetails()}},
\code{\link[=getOutcomeTable]{getOutcomeTable()}},
\code{\link[=getTarText]{getTarText()}},
\code{\link[=getTargetTable]{getTargetTable()}},
\code{\link[=kableDark]{kableDark()}},
\code{\link[=removeSpaces]{removeSpaces()}}
}
\concept{helper}
