% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimationQueries.R
\name{getCmTargets}
\alias{getCmTargets}
\title{A function to extract the targets found in cohort method}
\usage{
getCmTargets(
  connectionHandler,
  schema,
  cmTablePrefix = "cm_",
  cgTablePrefix = "cg_"
)
}
\arguments{
\item{connectionHandler}{A connection handler that connects to the database and extracts sql queries.  Create a connection handler via `ResultModelManager::ConnectionHandler$new()`.}

\item{schema}{The result database schema (e.g., 'main' for sqlite)}

\item{cmTablePrefix}{The prefix used for the cohort method results tables}

\item{cgTablePrefix}{The prefix used for the cohort generator results tables}
}
\value{
A data.frame with the cohort method target cohort ids and names.
}
\description{
A function to extract the targets found in cohort method
}
\details{
Specify the connectionHandler, the schema and the prefixes
}
\examples{
conDet <- getExampleConnectionDetails()

connectionHandler <- ResultModelManager::ConnectionHandler$new(conDet)

cohorts <- getCmTargets(
  connectionHandler = connectionHandler, 
  schema = 'main'
)

}
\seealso{
Other Estimation:
\code{\link[=.getCmVersion]{.getCmVersion()}},
\code{\link[=.getSccsVersion]{.getSccsVersion()}},
\code{\link[=getCMEstimation]{getCMEstimation()}},
\code{\link[=getCmDiagnosticsData]{getCmDiagnosticsData()}},
\code{\link[=getCmMetaEstimation]{getCmMetaEstimation()}},
\code{\link[=getCmNegativeControlEstimates]{getCmNegativeControlEstimates()}},
\code{\link[=getCmOutcomes]{getCmOutcomes()}},
\code{\link[=getCmPropensityModel]{getCmPropensityModel()}},
\code{\link[=getCmTable]{getCmTable()}},
\code{\link[=getSccsDiagnosticsData]{getSccsDiagnosticsData()}},
\code{\link[=getSccsEstimation]{getSccsEstimation()}},
\code{\link[=getSccsMetaEstimation]{getSccsMetaEstimation()}},
\code{\link[=getSccsModel]{getSccsModel()}},
\code{\link[=getSccsNegativeControlEstimates]{getSccsNegativeControlEstimates()}},
\code{\link[=getSccsOutcomes]{getSccsOutcomes()}},
\code{\link[=getSccsTable]{getSccsTable()}},
\code{\link[=getSccsTargets]{getSccsTargets()}},
\code{\link[=getSccsTimeToEvent]{getSccsTimeToEvent()}},
\code{\link[=plotCmEstimates]{plotCmEstimates()}},
\code{\link[=plotSccsEstimates]{plotSccsEstimates()}}
}
\concept{Estimation}
