% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortQueries.R
\name{getCohortInclusionStats}
\alias{getCohortInclusionStats}
\title{Extract the cohort inclusion stats}
\usage{
getCohortInclusionStats(
  connectionHandler,
  schema,
  cgTablePrefix = "cg_",
  databaseTable = "database_meta_data",
  cohortIds = NULL
)
}
\arguments{
\item{connectionHandler}{A connection handler that connects to the database and extracts sql queries.  Create a connection handler via `ResultModelManager::ConnectionHandler$new()`.}

\item{schema}{The result database schema (e.g., 'main' for sqlite)}

\item{cgTablePrefix}{The prefix used for the cohort generator results tables}

\item{databaseTable}{The name of the table with the database details (default 'database_meta_data')}

\item{cohortIds}{Optionally a list of cohortIds to restrict to}
}
\value{
Returns a data.frame with the cohort inclusion stats
}
\description{
This function extracts all cohort inclusion stats for the cohorts of interest.
}
\details{
Specify the connectionHandler, the schema and the cohort IDs
}
\examples{
conDet <- getExampleConnectionDetails()

connectionHandler <- ResultModelManager::ConnectionHandler$new(conDet)

cohortInclsuionsStats <- getCohortInclusionStats(
  connectionHandler = connectionHandler, 
  schema = 'main'
)

}
\seealso{
Other Cohorts: 
\code{\link{getCohortCounts}()},
\code{\link{getCohortDefinitions}()},
\code{\link{getCohortInclusionRules}()},
\code{\link{getCohortInclusionSummary}()},
\code{\link{getCohortMeta}()},
\code{\link{getCohortSubsetDefinitions}()},
\code{\link{getSubsetText}()},
\code{\link{processCohortDefinitionsForQuarto}()},
\code{\link{processCohorts}()},
\code{\link{restrictCohortDefinitionsForQuarto}()}
}
\concept{Cohorts}
