% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortQueries.R
\name{getCohortSubsetAttrition}
\alias{getCohortSubsetAttrition}
\title{Get cohort subset attrition}
\usage{
getCohortSubsetAttrition(
  connectionHandler,
  schema,
  cgTablePrefix = "cg_",
  databaseTable = "database_meta_data",
  cohortIds = NULL
)
}
\arguments{
\item{connectionHandler}{A connection handler object.}

\item{schema}{The database schema name.}

\item{cgTablePrefix}{Prefix for cohort generator tables. Default is 'cg_'.}

\item{databaseTable}{Name of the database metadata table. Default is 'database_meta_data'.}

\item{cohortIds}{Optional vector of cohort IDs to filter.}
}
\value{
A data.frame with attrition details for each cohort subset.
}
\description{
Retrieves attrition information for specified cohort subsets from the database.
}
\seealso{
Other Cohorts:
\code{\link[=getCohortAttrition]{getCohortAttrition()}},
\code{\link[=getCohortCounts]{getCohortCounts()}},
\code{\link[=getCohortDefinitions]{getCohortDefinitions()}},
\code{\link[=getCohortInclusionRules]{getCohortInclusionRules()}},
\code{\link[=getCohortInclusionStats]{getCohortInclusionStats()}},
\code{\link[=getCohortInclusionSummary]{getCohortInclusionSummary()}},
\code{\link[=getCohortMeta]{getCohortMeta()}},
\code{\link[=getCohortSubsetDefinitions]{getCohortSubsetDefinitions()}},
\code{\link[=getSubsetText]{getSubsetText()}},
\code{\link[=processCohortDefinitionsForQuarto]{processCohortDefinitionsForQuarto()}},
\code{\link[=processCohorts]{processCohorts()}},
\code{\link[=restrictCohortDefinitionsForQuarto]{restrictCohortDefinitionsForQuarto()}}
}
\concept{Cohorts}
