% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTargetInfo.R
\name{getParentTable}
\alias{getParentTable}
\title{Extract the parent cohort details from the target table}
\usage{
getParentTable(targetTable)
}
\arguments{
\item{targetTable}{The output from getTargetTable()}
}
\value{
Returns a data.frame with the parent cohorts used as targets
}
\description{
This function extracts the parents cohort details from the target table
}
\details{
Input the targetTable and this function extracts the parent cohorts and summary details about them
}
\examples{
conDet <- getExampleConnectionDetails()

connectionHandler <- ResultModelManager::ConnectionHandler$new(conDet)

targetTable <- getTargetTable(
  connectionHandler = connectionHandler, 
  schema = 'main'
)

parentTable <- getParentTable(targetTable)

}
\seealso{
Other Helpers: 
\code{\link{getTargetTable}()}
}
\concept{Helpers}
