% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionQueries.R
\name{getPredictionDiagnosticTable}
\alias{getPredictionDiagnosticTable}
\title{Extract specific diagnostic table}
\usage{
getPredictionDiagnosticTable(
  connectionHandler,
  schema,
  plpTablePrefix = "plp_",
  table = "diagnostic_participants",
  diagnosticId = NULL
)
}
\arguments{
\item{connectionHandler}{A connection handler that connects to the database and extracts sql queries.  Create a connection handler via `ResultModelManager::ConnectionHandler$new()`.}

\item{schema}{The result database schema (e.g., 'main' for sqlite)}

\item{plpTablePrefix}{The prefix used for the patient level prediction results tables}

\item{table}{The table to extract}

\item{diagnosticId}{(optional) restrict to the input diagnosticId}
}
\value{
Returns a data.frame with the specified table
}
\description{
This function extracts the specified diagnostic table
}
\details{
Specify the connectionHandler, the resultDatabaseSettings, the table of interest and (optionally) a diagnosticId to filter to
}
\examples{
conDet <- getExampleConnectionDetails()

connectionHandler <- ResultModelManager::ConnectionHandler$new(conDet)

diagPred <- getPredictionDiagnosticTable(
  connectionHandler = connectionHandler, 
  schema = 'main',
  table = 'diagnostic_predictors'
)

}
\seealso{
Other Prediction:
\code{\link[=getFullPredictionPerformances]{getFullPredictionPerformances()}},
\code{\link[=getPredictionAggregateTopPredictors]{getPredictionAggregateTopPredictors()}},
\code{\link[=getPredictionCohorts]{getPredictionCohorts()}},
\code{\link[=getPredictionCovariates]{getPredictionCovariates()}},
\code{\link[=getPredictionDiagnostics]{getPredictionDiagnostics()}},
\code{\link[=getPredictionHyperParamSearch]{getPredictionHyperParamSearch()}},
\code{\link[=getPredictionIntercept]{getPredictionIntercept()}},
\code{\link[=getPredictionLift]{getPredictionLift()}},
\code{\link[=getPredictionModelDesigns]{getPredictionModelDesigns()}},
\code{\link[=getPredictionOutcomes]{getPredictionOutcomes()}},
\code{\link[=getPredictionPerformanceTable]{getPredictionPerformanceTable()}},
\code{\link[=getPredictionPerformances]{getPredictionPerformances()}},
\code{\link[=getPredictionTargets]{getPredictionTargets()}},
\code{\link[=getPredictionTopPredictors]{getPredictionTopPredictors()}}
}
\concept{Prediction}
