% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SccQueries.R
\name{getSccAnalysisSettings}
\alias{getSccAnalysisSettings}
\title{Extract the self controlled cohort analysis settings}
\usage{
getSccAnalysisSettings(connectionHandler, schema, sccTablePrefix = "scc_")
}
\arguments{
\item{connectionHandler}{A connection handler that connects to the database and extracts sql queries.  Create a connection handler via `ResultModelManager::ConnectionHandler$new()`.}

\item{schema}{The result database schema (e.g., 'main' for sqlite)}

\item{sccTablePrefix}{The prefix used for the self controlled cohort results tables}
}
\value{
A data.frame with the analysis settings
}
\description{
Extract the self controlled cohort analysis settings
}
\details{
Returns the analysis settings used by the SelfControlledCohort analyses
}
\seealso{
Other SelfControlledCohort: 
\code{\link{createSccSignalsTable}()},
\code{\link{getSccDiagnosticsData}()},
\code{\link{getSccEstimation}()},
\code{\link{getSccMetaEstimation}()},
\code{\link{getSccNegativeControlEstimates}()},
\code{\link{getSccOutcomes}()},
\code{\link{getSccSignals}()},
\code{\link{getSccSummaryStats}()},
\code{\link{getSccTargets}()},
\code{\link{plotSccBoxPlot}()},
\code{\link{plotSccForest}()},
\code{\link{plotSccSystematicError}()}
}
\concept{SelfControlledCohort}
