% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SccQueries.R
\name{getSccDiagnosticsData}
\alias{getSccDiagnosticsData}
\title{Extract the self controlled cohort diagnostic summary}
\usage{
getSccDiagnosticsData(
  connectionHandler,
  schema,
  sccTablePrefix = "scc_",
  cgTablePrefix = "cg_",
  databaseTable = "database_meta_data",
  analysisIds = NULL,
  targetIds = NULL,
  outcomeIds = NULL
)
}
\arguments{
\item{connectionHandler}{A connection handler that connects to the database and extracts sql queries.  Create a connection handler via `ResultModelManager::ConnectionHandler$new()`.}

\item{schema}{The result database schema (e.g., 'main' for sqlite)}

\item{sccTablePrefix}{The prefix used for the self controlled cohort results tables}

\item{cgTablePrefix}{The prefix used for the cohort generator results tables}

\item{databaseTable}{The name of the table with the database details (default 'database_meta_data')}

\item{analysisIds}{A vector of analysis ids to restrict the results to}

\item{targetIds}{A vector of target cohort ids to restrict the results to}

\item{outcomeIds}{A vector of outcome cohort ids to restrict the results to}
}
\value{
Returns a data.frame with the columns:
\itemize{
 \item{databaseName the database name}
 \item{databaseId the unique identifier for the database}
 \item{analysisId the analysis unique identifier}
 \item{description an analysis description}
 \item{targetId the target cohort id}
 \item{targetName the target cohort name}
 \item{outcomeId the outcome cohort id}
 \item{outcomeName the outcome cohort name}
 \item{unblind whether the pair is unblinded}
 \item{diagnostics a list column of the diagnostic values}
 \item{summaryValue a summary of the diagnostic pass/fail status}
 }
}
\description{
Returns the per database diagnostic summary for the selected exposure
(target) and outcome cohort pairs
}
\details{
Specify the connectionHandler, the schema and the optional target/outcome/analysis
cohort identifiers to restrict the results
}
\seealso{
Other SelfControlledCohort: 
\code{\link{createSccSignalsTable}()},
\code{\link{getSccAnalysisSettings}()},
\code{\link{getSccEstimation}()},
\code{\link{getSccMetaEstimation}()},
\code{\link{getSccNegativeControlEstimates}()},
\code{\link{getSccOutcomes}()},
\code{\link{getSccSignals}()},
\code{\link{getSccSummaryStats}()},
\code{\link{getSccTargets}()},
\code{\link{plotSccBoxPlot}()},
\code{\link{plotSccForest}()},
\code{\link{plotSccSystematicError}()}
}
\concept{SelfControlledCohort}
