% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SccQueries.R
\name{getSccEstimation}
\alias{getSccEstimation}
\title{Extract the self controlled cohort effect estimates per database}
\usage{
getSccEstimation(
  connectionHandler,
  schema,
  sccTablePrefix = "scc_",
  cgTablePrefix = "cg_",
  databaseTable = "database_meta_data",
  analysisIds = NULL,
  targetIds = NULL,
  outcomeIds = NULL
)
}
\arguments{
\item{connectionHandler}{A connection handler that connects to the database and extracts sql queries.  Create a connection handler via `ResultModelManager::ConnectionHandler$new()`.}

\item{schema}{The result database schema (e.g., 'main' for sqlite)}

\item{sccTablePrefix}{The prefix used for the self controlled cohort results tables}

\item{cgTablePrefix}{The prefix used for the cohort generator results tables}

\item{databaseTable}{The name of the table with the database details (default 'database_meta_data')}

\item{analysisIds}{A vector of analysis ids to restrict the results to}

\item{targetIds}{A vector of target cohort ids to restrict the results to}

\item{outcomeIds}{A vector of outcome cohort ids to restrict the results to}
}
\value{

}
\description{
Returns the per database incidence rate ratio estimates for the selected
exposure (target) and outcome cohort pairs.  Effect estimates are only
returned for exposure-outcome pairs that have been unblinded - the blinding
status is taken from the UNBLIND diagnostic stored in the
scc_diagnostics_summary table
}
\details{
Specify the connectionHandler, the schema and the optional target/outcome/analysis
cohort identifiers to restrict the results
}
\seealso{
Other SelfControlledCohort: 
\code{\link{createSccSignalsTable}()},
\code{\link{getSccAnalysisSettings}()},
\code{\link{getSccDiagnosticsData}()},
\code{\link{getSccMetaEstimation}()},
\code{\link{getSccNegativeControlEstimates}()},
\code{\link{getSccOutcomes}()},
\code{\link{getSccSignals}()},
\code{\link{getSccSummaryStats}()},
\code{\link{getSccTargets}()},
\code{\link{plotSccBoxPlot}()},
\code{\link{plotSccForest}()},
\code{\link{plotSccSystematicError}()}
}
\concept{SelfControlledCohort}
