% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SccQueries.R
\name{getSccNegativeControlEstimates}
\alias{getSccNegativeControlEstimates}
\title{Extract the negative control estimates for the self controlled cohort analyses}
\usage{
getSccNegativeControlEstimates(
  connectionHandler,
  schema,
  sccTablePrefix = "scc_",
  cgTablePrefix = "cg_",
  databaseTable = "database_meta_data",
  analysisIds = NULL,
  targetIds = NULL,
  outcomeIds = NULL,
  databaseIds = NULL
)
}
\arguments{
\item{connectionHandler}{A connection handler that connects to the database and extracts sql queries.  Create a connection handler via `ResultModelManager::ConnectionHandler$new()`.}

\item{schema}{The result database schema (e.g., 'main' for sqlite)}

\item{sccTablePrefix}{The prefix used for the self controlled cohort results tables}

\item{cgTablePrefix}{The prefix used for the cohort generator results tables}

\item{databaseTable}{The name of the table with the database details (default 'database_meta_data')}

\item{analysisIds}{A vector of analysis ids to restrict the results to}

\item{targetIds}{A vector of target cohort ids to restrict the results to}

\item{outcomeIds}{A vector of outcome cohort ids to restrict the results to}

\item{databaseIds}{A vector of database ids to restrict the results to}
}
\value{
A data.frame with the negative control estimates
}
\description{
Returns the estimates for exposure-outcome pairs that are negative controls
}
\details{
Specify the connectionHandler, the schema and the target/outcome identifiers.
Negative controls are identified via the scc_outcome_exposure table (pairs
with a non missing true effect size)
}
\seealso{
Other SelfControlledCohort: 
\code{\link{createSccSignalsTable}()},
\code{\link{getSccAnalysisSettings}()},
\code{\link{getSccDiagnosticsData}()},
\code{\link{getSccEstimation}()},
\code{\link{getSccMetaEstimation}()},
\code{\link{getSccOutcomes}()},
\code{\link{getSccSignals}()},
\code{\link{getSccSummaryStats}()},
\code{\link{getSccTargets}()},
\code{\link{plotSccBoxPlot}()},
\code{\link{plotSccForest}()},
\code{\link{plotSccSystematicError}()}
}
\concept{SelfControlledCohort}
