% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SccQueries.R
\name{getSccSummaryStats}
\alias{getSccSummaryStats}
\title{Extract the self controlled cohort summary statistics for boxplots}
\usage{
getSccSummaryStats(
  connectionHandler,
  schema,
  sccTablePrefix = "scc_",
  cgTablePrefix = "cg_",
  databaseTable = "database_meta_data",
  statTypes = NULL,
  analysisIds = NULL,
  targetIds = NULL,
  outcomeIds = NULL
)
}
\arguments{
\item{connectionHandler}{A connection handler that connects to the database and extracts sql queries.  Create a connection handler via `ResultModelManager::ConnectionHandler$new()`.}

\item{schema}{The result database schema (e.g., 'main' for sqlite)}

\item{sccTablePrefix}{The prefix used for the self controlled cohort results tables}

\item{cgTablePrefix}{The prefix used for the cohort generator results tables}

\item{databaseTable}{The name of the table with the database details (default 'database_meta_data')}

\item{statTypes}{A vector of stat types to restrict the results to
(e.g. time_exposed, time_to_outcome, time_to_outcome_exposed,
time_to_outcome_unexposed)}

\item{analysisIds}{A vector of analysis ids to restrict the results to}

\item{targetIds}{A vector of target cohort ids to restrict the results to}

\item{outcomeIds}{A vector of outcome cohort ids to restrict the results to}
}
\value{
Returns a data.frame with the summary statistics per database
}
\description{
Returns the distribution statistics (e.g. time at risk, time to outcome)
used to create boxplots for the selected exposure and outcome cohorts
}
\details{
Specify the connectionHandler, the schema and the optional stat type,
target/outcome/analysis identifiers to restrict the results
}
\seealso{
Other SelfControlledCohort: 
\code{\link{createSccSignalsTable}()},
\code{\link{getSccAnalysisSettings}()},
\code{\link{getSccDiagnosticsData}()},
\code{\link{getSccEstimation}()},
\code{\link{getSccMetaEstimation}()},
\code{\link{getSccNegativeControlEstimates}()},
\code{\link{getSccOutcomes}()},
\code{\link{getSccSignals}()},
\code{\link{getSccTargets}()},
\code{\link{plotSccBoxPlot}()},
\code{\link{plotSccForest}()},
\code{\link{plotSccSystematicError}()}
}
\concept{SelfControlledCohort}
