% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{getTarText}
\alias{getTarText}
\title{output a string summarizing the time at risk}
\usage{
getTarText(tar)
}
\arguments{
\item{tar}{a list containing the four tar setting}
}
\value{
An character string describing the tar
}
\description{
This returns a character that describes the time at risk
}
\details{
Input the tar settings
}
\examples{
targetText <- getTarText(
list(
startAnchor = 'cohort_start',
riskWindowStart = 1,
endAnchor = 'cohort_end',
riskWindowEnd = 0
)
)

}
\seealso{
Other helper:
\code{\link[=addTarColumn]{addTarColumn()}},
\code{\link[=formatBinaryCovariateName]{formatBinaryCovariateName()}},
\code{\link[=formatTargetPopulation]{formatTargetPopulation()}},
\code{\link[=getExampleConnectionDetails]{getExampleConnectionDetails()}},
\code{\link[=getOutcomeTable]{getOutcomeTable()}},
\code{\link[=getTargetTable]{getTargetTable()}},
\code{\link[=kableDark]{kableDark()}},
\code{\link[=removeSpaces]{removeSpaces()}}
}
\concept{helper}
