% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OhdsiReportGenerator.R
\docType{package}
\name{OhdsiReportGenerator-package}
\alias{OhdsiReportGenerator}
\alias{OhdsiReportGenerator-package}
\title{OhdsiReportGenerator: Extract results and generate plots, presentations and reports from the Observational Health Data Sciences and Informatics results database}
\description{
Extract analysis results into R data.frames from the Observational Health Data Sciences and Informatics result database and generate reports that summarize results in HTML format.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/OhdsiReportGenerator}
  \item \url{https://github.com/OHDSI/OhdsiReportGenerator}
  \item Report bugs at \url{https://github.com/OHDSI/OhdsiReportGenerator/issues}
}

}
\author{
\strong{Maintainer}: Jenna Reps \email{jreps@its.jnj.com}

Authors:
\itemize{
  \item Anthony Sena \email{...@its.jnj.com}
}

}
\keyword{internal}
