% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SccPlots.R
\name{plotSccBoxPlot}
\alias{plotSccBoxPlot}
\title{Plot the self controlled cohort summary statistics as boxplots}
\usage{
plotSccBoxPlot(data, xLabel = "Data source", yLabel = "Days")
}
\arguments{
\item{data}{A data.frame with the columns databaseName, minimum, p25,
median, p75, maximum (and optionally mean, sd, p10, p90)}

\item{xLabel}{The label for the x axis}

\item{yLabel}{The label for the y axis}
}
\value{
A ggplot object
}
\description{
Create boxplots of the summary statistics (e.g. time at risk, time to
outcome) returned by \code{\link{getSccSummaryStats}}
}
\seealso{
Other SelfControlledCohort: 
\code{\link{createSccSignalsTable}()},
\code{\link{getSccAnalysisSettings}()},
\code{\link{getSccDiagnosticsData}()},
\code{\link{getSccEstimation}()},
\code{\link{getSccMetaEstimation}()},
\code{\link{getSccNegativeControlEstimates}()},
\code{\link{getSccOutcomes}()},
\code{\link{getSccSignals}()},
\code{\link{getSccSummaryStats}()},
\code{\link{getSccTargets}()},
\code{\link{plotSccForest}()},
\code{\link{plotSccSystematicError}()}
}
\concept{SelfControlledCohort}
