% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimationPlots.R
\name{plotSccsEstimates}
\alias{plotSccsEstimates}
\title{Plots the self controlled case series results for one analysis}
\usage{
plotSccsEstimates(
  sccsData,
  sccsDiagnostics = NULL,
  sccsMeta = NULL,
  includeCounts = TRUE,
  selectedAnalysisId
)
}
\arguments{
\item{sccsData}{The self controlled case series data}

\item{sccsDiagnostics}{(optional) The self controlled case series diagnostic data}

\item{sccsMeta}{(optional) The self controlled case series evidence synthesis data}

\item{includeCounts}{Whether to include count on the plot}

\item{selectedAnalysisId}{The analysis ID of interest to plot}
}
\value{
Returns a ggplot with the estimates
}
\description{
Creates nice self controlled case series plots
}
\details{
Input the self controlled case series data
}
\examples{

conDet <- getExampleConnectionDetails()

connectionHandler <- ResultModelManager::ConnectionHandler$new(conDet)

sccsEst <- getSccsEstimation(
  connectionHandler = connectionHandler, 
  schema = 'main',
  targetIds = 1,
  outcomeIds = 3
)
plotSccsEstimates(
  sccsData = sccsEst, 
  sccsMeta = NULL, 
  selectedAnalysisId = 1
)

}
\seealso{
Other Estimation:
\code{\link[=.getCmVersion]{.getCmVersion()}},
\code{\link[=.getSccsVersion]{.getSccsVersion()}},
\code{\link[=getCMEstimation]{getCMEstimation()}},
\code{\link[=getCmDiagnosticsData]{getCmDiagnosticsData()}},
\code{\link[=getCmMetaEstimation]{getCmMetaEstimation()}},
\code{\link[=getCmNegativeControlEstimates]{getCmNegativeControlEstimates()}},
\code{\link[=getCmOutcomes]{getCmOutcomes()}},
\code{\link[=getCmPropensityModel]{getCmPropensityModel()}},
\code{\link[=getCmTable]{getCmTable()}},
\code{\link[=getCmTargets]{getCmTargets()}},
\code{\link[=getSccsDiagnosticsData]{getSccsDiagnosticsData()}},
\code{\link[=getSccsEstimation]{getSccsEstimation()}},
\code{\link[=getSccsMetaEstimation]{getSccsMetaEstimation()}},
\code{\link[=getSccsModel]{getSccsModel()}},
\code{\link[=getSccsNegativeControlEstimates]{getSccsNegativeControlEstimates()}},
\code{\link[=getSccsOutcomes]{getSccsOutcomes()}},
\code{\link[=getSccsTable]{getSccsTable()}},
\code{\link[=getSccsTargets]{getSccsTargets()}},
\code{\link[=getSccsTimeToEvent]{getSccsTimeToEvent()}},
\code{\link[=plotCmEstimates]{plotCmEstimates()}}
}
\concept{Estimation}
