% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortQueries.R
\name{restrictCohortDefinitionsForQuarto}
\alias{restrictCohortDefinitionsForQuarto}
\title{Function to figure out the target, comparator, outcome and indication ids of interest
for the quarto report based on the user inputs}
\usage{
restrictCohortDefinitionsForQuarto(
  connectionHandler,
  schema,
  cohortDefinitions,
  targetId,
  outcomeIds,
  comparatorIds,
  restrictTargetToIndications,
  indicationIds,
  includeCohortMethod
)
}
\arguments{
\item{connectionHandler}{A connection handler that connects to the database and extracts sql queries.  Create a connection handler via `ResultModelManager::ConnectionHandler$new()`.}

\item{schema}{The result database schema (e.g., 'main' for sqlite)}

\item{cohortDefinitions}{The output of `processCohortDefinitionsForQuarto()`}

\item{targetId}{a parent target id}

\item{outcomeIds}{a vector of outcome ids}

\item{comparatorIds}{NULL or a vector of comparator parent ids}

\item{restrictTargetToIndications}{whether to restrict the target ids to cohorts nested in indicationIds}

\item{indicationIds}{The indication ids of interest for restrictTargetToIndications}

\item{includeCohortMethod}{If TRUE, when comparatorIds is NULL all comparators included in CohortMethod are included in the report}
}
\value{
A list of targetIdsOfInterest that is a vector of cohortIds that are targets of interest to include in the report,
comparatorIds a vector of cohortIds that are comparators, 
comparatorIdsOfInterest a vector of cohortIds that are comparators of interest, 
outcomeIdsOfInterest  a vector of cohortIds that are outcomes of interest and
indicationIdsOfInterest  a vector of cohortIds that are indications of interest.
}
\description{
Function to figure out the target, comparator, outcome and indication ids of interest
for the quarto report based on the user inputs
}
\details{
This function finds the targets, comparators, indications and outcomes of interest based
on the user inputs for quarto report generation.
}
\seealso{
Other Cohorts: 
\code{\link{getCohortCounts}()},
\code{\link{getCohortDefinitions}()},
\code{\link{getCohortInclusionRules}()},
\code{\link{getCohortInclusionStats}()},
\code{\link{getCohortInclusionSummary}()},
\code{\link{getCohortMeta}()},
\code{\link{getCohortSubsetDefinitions}()},
\code{\link{getSubsetText}()},
\code{\link{processCohortDefinitionsForQuarto}()},
\code{\link{processCohorts}()}
}
\concept{Cohorts}
