% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Indexes.R
\name{createCohortIndexes}
\alias{createCohortIndexes}
\title{A function to add indexes to the Cohort Generator results}
\usage{
createCohortIndexes(connectionHandler, schema, cgTablePrefix = "cg_")
}
\arguments{
\item{connectionHandler}{A connection handler that connects to the database and extracts sql queries.  Create a connection handler via `ResultModelManager::ConnectionHandler$new()`.}

\item{schema}{The result database schema (e.g., 'main' for sqlite)}

\item{cgTablePrefix}{The prefix used for the cohort generator results tables}
}
\description{
A function to add indexes to the Cohort Generator results
}
\details{
Specify the connectionHandler, the schema and the prefixes
}
\seealso{
Other Indexes: 
\code{\link{createCharacterizationIndexes}()},
\code{\link{createIncidenceIndexes}()},
\code{\link{createSccsIndexes}()}
}
\concept{Indexes}
