% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Indexes.R
\name{createIncidenceIndexes}
\alias{createIncidenceIndexes}
\title{A function to add indexes to the incidence results}
\usage{
createIncidenceIndexes(connectionHandler, schema, ciTablePrefix = "ci_")
}
\arguments{
\item{connectionHandler}{A connection handler that connects to the database and extracts sql queries.  Create a connection handler via `ResultModelManager::ConnectionHandler$new()`.}

\item{schema}{The result database schema (e.g., 'main' for sqlite)}

\item{ciTablePrefix}{The prefix used for the cohort incidence results tables}
}
\description{
A function to add indexes to the incidence results
}
\details{
Specify the connectionHandler, the schema and the prefixes
}
\seealso{
Other Indexes: 
\code{\link{createCharacterizationIndexes}()},
\code{\link{createCohortIndexes}()},
\code{\link{createSccsIndexes}()}
}
\concept{Indexes}
