% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Indexes.R
\name{createSccsIndexes}
\alias{createSccsIndexes}
\title{A function to add indexes to the SCCS results}
\usage{
createSccsIndexes(connectionHandler, schema, sccsTablePrefix = "sccs_")
}
\arguments{
\item{connectionHandler}{A connection handler that connects to the database and extracts sql queries.  Create a connection handler via `ResultModelManager::ConnectionHandler$new()`.}

\item{schema}{The result database schema (e.g., 'main' for sqlite)}

\item{sccsTablePrefix}{The prefix used for the cohort generator results tables}
}
\description{
A function to add indexes to the SCCS results
}
\details{
Specify the connectionHandler, the schema and the prefixes
}
\seealso{
Other Indexes: 
\code{\link{createCharacterizationIndexes}()},
\code{\link{createCohortIndexes}()},
\code{\link{createIncidenceIndexes}()}
}
\concept{Indexes}
