% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortQueries.R
\name{.getCgVersion}
\alias{.getCgVersion}
\title{An internal function to determine the version of CohortGenerator is 
used to store results}
\usage{
.getCgVersion(connectionHandler, schema, cgTablePrefix = "cg_")
}
\arguments{
\item{connectionHandler}{A connection handler that connects to the database and extracts sql queries.  Create a connection handler via `ResultModelManager::ConnectionHandler$new()`.}

\item{schema}{The result database schema (e.g., 'main' for sqlite)}

\item{cmTablePrefix}{The prefix used for the cohort method results tables}
}
\value{
A integer with the major version number of CohortGenerator
}
\description{
An internal function to determine the version of CohortGenerator is 
used to store results
}
\details{
Specify the connectionHandler, the schema and the prefixes. This
query will attempt to identify if CohortGenerator v0.x was used by 
inspecing the migration table. When the migration_order is >= 3
then v1 of CohortGenerator was used.
}
\examples{
conDet <- getExampleConnectionDetails()

connectionHandler <- ResultModelManager::ConnectionHandler$new(conDet)

version <- .getCgVersion(
  connectionHandler = connectionHandler, 
  schema = 'main'
)

}
\seealso{
Other Estimation: 
\code{\link{getCMEstimation}()},
\code{\link{getCmDiagnosticsData}()},
\code{\link{getCmMetaEstimation}()},
\code{\link{getCmNegativeControlEstimates}()},
\code{\link{getCmOutcomes}()},
\code{\link{getCmPropensityModel}()},
\code{\link{getCmTable}()},
\code{\link{getCmTargets}()},
\code{\link{getSccsDiagnosticsData}()},
\code{\link{getSccsEstimation}()},
\code{\link{getSccsMetaEstimation}()},
\code{\link{getSccsModel}()},
\code{\link{getSccsNegativeControlEstimates}()},
\code{\link{getSccsOutcomes}()},
\code{\link{getSccsTable}()},
\code{\link{getSccsTargets}()},
\code{\link{getSccsTimeToEvent}()},
\code{\link{plotCmEstimates}()},
\code{\link{plotSccsEstimates}()}
}
\concept{Estimation}
