% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimationQueries.R
\name{.getSccsVersion}
\alias{.getSccsVersion}
\title{An internal function to determine the version of SCCS used 
to store results}
\usage{
.getSccsVersion(connectionHandler, schema, sccsTablePrefix = "sccs_")
}
\arguments{
\item{connectionHandler}{A connection handler that connects to the database and extracts sql queries.  Create a connection handler via `ResultModelManager::ConnectionHandler$new()`.}

\item{schema}{The result database schema (e.g., 'main' for sqlite)}

\item{sccsTablePrefix}{The prefix used for the cohort generator results tables}
}
\value{
A integer with the major version number of cohort method
}
\description{
An internal function to determine the version of SCCS used 
to store results
}
\details{
Specify the connectionHandler, the schema and the prefixes. This
query will attempt to identify if SelfControlledCaseSeries v6.1.4 was used by 
inspecing the migration table. When the migration_order is >= 3
then v6.1.4 of SelfControlledCaseSeries was used.
}
\seealso{
Other Estimation: 
\code{\link{.getCmVersion}()},
\code{\link{getCMEstimation}()},
\code{\link{getCmDiagnosticsData}()},
\code{\link{getCmMetaEstimation}()},
\code{\link{getCmNegativeControlEstimates}()},
\code{\link{getCmOutcomes}()},
\code{\link{getCmPropensityModel}()},
\code{\link{getCmTable}()},
\code{\link{getCmTargets}()},
\code{\link{getSccsDiagnosticsData}()},
\code{\link{getSccsEstimation}()},
\code{\link{getSccsMetaEstimation}()},
\code{\link{getSccsModel}()},
\code{\link{getSccsNegativeControlEstimates}()},
\code{\link{getSccsOutcomes}()},
\code{\link{getSccsTable}()},
\code{\link{getSccsTargets}()},
\code{\link{getSccsTimeToEvent}()},
\code{\link{plotCmEstimates}()},
\code{\link{plotSccsEstimates}()}
}
\concept{Estimation}
