% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{formatBinaryCovariateName}
\alias{formatBinaryCovariateName}
\title{formatBinaryCovariateName}
\usage{
formatBinaryCovariateName(data)
}
\arguments{
\item{data}{The data.frame with the covariateName column}
}
\value{
The data data.frame object with the ovariateName column changed to be more friendly
}
\description{
Removes the long part of the covariate name to make it friendly
}
\details{
Makes the covariateName more friendly and shorter
}
\examples{
formatBinaryCovariateName(data.frame(
covariateName = c("fdfgfgf: dgdgff","made up test")
))

}
\seealso{
Other helper: 
\code{\link{addTarColumn}()},
\code{\link{getExampleConnectionDetails}()},
\code{\link{kableDark}()},
\code{\link{printReactable}()},
\code{\link{removeSpaces}()}
}
\concept{helper}
