% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharacterzationQueries.R
\name{getBinaryCaseSeries}
\alias{getBinaryCaseSeries}
\title{A function to extract case series characterization results}
\usage{
getBinaryCaseSeries(
  connectionHandler,
  schema,
  cTablePrefix = "c_",
  cgTablePrefix = "cg_",
  databaseTable = "database_meta_data",
  characterizationTargetId = NULL,
  characterizationCaseId = NULL,
  outcomeId = NULL,
  databaseIds = NULL,
  riskWindowStart = NULL,
  riskWindowEnd = NULL,
  startAnchor = NULL,
  endAnchor = NULL,
  conceptIds = NULL,
  minVal = NULL
)
}
\arguments{
\item{connectionHandler}{A connection handler that connects to the database and extracts sql queries.  Create a connection handler via `ResultModelManager::ConnectionHandler$new()`.}

\item{schema}{The result database schema (e.g., 'main' for sqlite)}

\item{cTablePrefix}{The prefix used for the characterization results tables}

\item{cgTablePrefix}{The prefix used for the cohort generator results tables}

\item{databaseTable}{The name of the table with the database details (default 'database_meta_data')}

\item{characterizationTargetId}{The characterization target id to restrict results to}

\item{characterizationCaseId}{The characterization case id to restrict results to}

\item{outcomeId}{Am integer corresponding to the outcome cohort ID}

\item{databaseIds}{(optional) One or more unique identifiers for the databases}

\item{riskWindowStart}{(optional) A riskWindowStart to restrict to}

\item{riskWindowEnd}{(optional) A riskWindowEnd to restrict to}

\item{startAnchor}{(optional) A startAnchor to restrict to}

\item{endAnchor}{(optional) An endAnchor to restrict to}

\item{conceptIds}{(optional) An conceptIds to restrict to}

\item{minVal}{(optional) the minimum averageVal to extract}
}
\value{
A data.frame with the characterization case series results
}
\description{
A function to extract case series characterization results
}
\details{
Specify the connectionHandler, the schema and the target/outcome cohort IDs
}
\examples{
conDet <- getExampleConnectionDetails()

connectionHandler <- ResultModelManager::ConnectionHandler$new(conDet)

cs <- getBinaryCaseSeries(
  connectionHandler = connectionHandler, 
  schema = 'main',
  characterizationTargetId = 1, 
  outcomeId = 3
)

}
\seealso{
Other Characterization: 
\code{\link{characterizationCompareBinary}()},
\code{\link{characterizationCompareContinuous}()},
\code{\link{getBinaryRiskFactors}()},
\code{\link{getBinaryTargetBaseline}()},
\code{\link{getCaseCounts}()},
\code{\link{getCharacterizationCaseSettings}()},
\code{\link{getCharacterizationDemographics}()},
\code{\link{getCharacterizationTargetSettings}()},
\code{\link{getContinuousCaseSeries}()},
\code{\link{getContinuousRiskFactors}()},
\code{\link{getContinuousTargetBaseline}()},
\code{\link{getDechallengeRechallenge}()},
\code{\link{getDechallengeRechallengeFails}()},
\code{\link{getIncidenceRates}()},
\code{\link{getIncidenceTargetSettings}()},
\code{\link{getNonCaseCounts}()},
\code{\link{getOutcomesUsedInCharacterization}()},
\code{\link{getOutcomesUsedInIncidence}()},
\code{\link{getTargetCounts}()},
\code{\link{getTargetsUsedInCharacterization}()},
\code{\link{getTargetsUsedInIncidence}()},
\code{\link{getTimeToEvent}()},
\code{\link{plotAgeDistributions}()},
\code{\link{plotSexDistributions}()},
\code{\link{viewIncidenceRate}()}
}
\concept{Characterization}
