% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharacterzationQueries.R
\name{getBinaryTargetBaseline}
\alias{getBinaryTargetBaseline}
\title{Extract the aggregate covariates for the target ids of interest}
\usage{
getBinaryTargetBaseline(
  connectionHandler,
  schema,
  cTablePrefix = "c_",
  cgTablePrefix = "cg_",
  databaseTable = "database_meta_data",
  characterizationTargetIds = NULL,
  analysisIds = NULL,
  covariateIds = NULL,
  conceptIds = NULL,
  databaseIds = NULL,
  includeNames = TRUE,
  minThreshold = NULL
)
}
\arguments{
\item{connectionHandler}{A connection handler that connects to the database and extracts sql queries.  Create a connection handler via `ResultModelManager::ConnectionHandler$new()`.}

\item{schema}{The result database schema (e.g., 'main' for sqlite)}

\item{cTablePrefix}{The prefix used for the characterization results tables}

\item{cgTablePrefix}{The prefix used for the cohort generator results tables}

\item{databaseTable}{The name of the table with the database details (default 'database_meta_data')}

\item{characterizationTargetIds}{The characterization target cohort ids of interest}

\item{analysisIds}{The analysisIds of the covariate to restrict results to}

\item{covariateIds}{The covariateIds to restict results to}

\item{conceptIds}{The conceptIds of the covariate to restrict results to}

\item{databaseIds}{The databaseIds of the covariate to restrict results to}

\item{includeNames}{Whether to add database and cohort names (setting to FALSE will make extraction quicker)}

\item{minThreshold}{(optional) The minimum average value for results to be returned}
}
\value{
Returns a data.frame with the columns:
\itemize{
 \item{databaseName the name of the database}
 \item{databaseId the unique identifier of the database}
 \item{targetName the target cohort name}
 \item{targetId the target cohort unique identifier}
 \item{minPriorObservation the }
 \item{limitToFirstINDays the }
 \item{covariateName the }
 \item{covariateId the }
 \item{analysisId the }
 \item{sumValue the }
 \item{averageValue the }
 }
}
\description{
This function extracts the specified covariates for the specified targets
}
\details{
Specify the connectionHandler, the schema and the target cohort IDs
}
\examples{
conDet <- getExampleConnectionDetails()

connectionHandler <- ResultModelManager::ConnectionHandler$new(conDet)

btb <- getBinaryTargetBaseline(
 connectionHandler = connectionHandler, 
 schema = 'main', 
 characterizationTargetIds = 1
)
 
}
\seealso{
Other Characterization:
\code{\link[=characterizationCompareBinary]{characterizationCompareBinary()}},
\code{\link[=characterizationCompareContinuous]{characterizationCompareContinuous()}},
\code{\link[=getBinaryCaseSeries]{getBinaryCaseSeries()}},
\code{\link[=getBinaryRiskFactors]{getBinaryRiskFactors()}},
\code{\link[=getCaseCounts]{getCaseCounts()}},
\code{\link[=getCharacterizationCaseSettings]{getCharacterizationCaseSettings()}},
\code{\link[=getCharacterizationDemographics]{getCharacterizationDemographics()}},
\code{\link[=getCharacterizationTargetSettings]{getCharacterizationTargetSettings()}},
\code{\link[=getContinuousCaseSeries]{getContinuousCaseSeries()}},
\code{\link[=getContinuousRiskFactors]{getContinuousRiskFactors()}},
\code{\link[=getContinuousTargetBaseline]{getContinuousTargetBaseline()}},
\code{\link[=getDechallengeRechallenge]{getDechallengeRechallenge()}},
\code{\link[=getDechallengeRechallengeFails]{getDechallengeRechallengeFails()}},
\code{\link[=getIncidenceRates]{getIncidenceRates()}},
\code{\link[=getIncidenceTargetSettings]{getIncidenceTargetSettings()}},
\code{\link[=getNonCaseCounts]{getNonCaseCounts()}},
\code{\link[=getOutcomesUsedInCharacterization]{getOutcomesUsedInCharacterization()}},
\code{\link[=getOutcomesUsedInIncidence]{getOutcomesUsedInIncidence()}},
\code{\link[=getTargetCounts]{getTargetCounts()}},
\code{\link[=getTargetsUsedInCharacterization]{getTargetsUsedInCharacterization()}},
\code{\link[=getTargetsUsedInIncidence]{getTargetsUsedInIncidence()}},
\code{\link[=getTimeToEvent]{getTimeToEvent()}},
\code{\link[=plotAgeDistributions]{plotAgeDistributions()}},
\code{\link[=plotSexDistributions]{plotSexDistributions()}},
\code{\link[=viewIncidenceRate]{viewIncidenceRate()}}
}
\concept{Characterization}
