% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharacterzationQueries.R
\name{getCaseCounts}
\alias{getCaseCounts}
\title{Extract the outcome cohort counts result}
\usage{
getCaseCounts(
  connectionHandler,
  schema,
  cTablePrefix = "c_",
  cgTablePrefix = "cg_",
  databaseTable = "database_meta_data",
  targetIds = NULL,
  outcomeIds = NULL
)
}
\arguments{
\item{connectionHandler}{A connection handler that connects to the database and extracts sql queries.  Create a connection handler via `ResultModelManager::ConnectionHandler$new()`.}

\item{schema}{The result database schema (e.g., 'main' for sqlite)}

\item{cTablePrefix}{The prefix used for the characterization results tables}

\item{cgTablePrefix}{The prefix used for the cohort generator results tables}

\item{databaseTable}{The name of the table with the database details (default 'database_meta_data')}

\item{targetIds}{A vector of integers corresponding to the target cohort IDs}

\item{outcomeIds}{A vector of integers corresponding to the outcome cohort IDs}
}
\value{
Returns a data.frame with the columns:
\itemize{
 \item{databaseName} \description{the name of the database}
 \item{targetName} \description{the target cohort name}
 \item{targetId} \description{the target cohort unique identifier}
 \item{outcomeName} \description{the outcome name}
 \item{outcomeId} \description{the outcome unique identifier}
 \item{rowCount} \description{the number of entries in the cohort}
 \item{personCount} \description{the number of people in the cohort}
 \item{minPriorObservation} \description{the minimum required observation days prior to index for an entry}
 \item{outcomeWashoutDays} \description{patients with the outcome occurring within this number of days prior to index are excluded (NA means no exclusion)}
\item{riskWindowStart} \description{the number of days ofset the start anchor that is the start of the time-at-risk}
\item{startAnchor} \description{the start anchor is either the target cohort start or cohort end date}
\item{riskWindowEnd} \description{the number of days ofset the end anchor that is the end of the time-at-risk}
\item{endAnchor} \description{the end anchor is either the target cohort start or cohort end date}
}
}
\description{
This function extracts outcome cohort counts across databases in the results for specified target and outcome cohorts.
}
\details{
Specify the connectionHandler, the schema and the target/outcome cohort IDs
}
\seealso{
Other {Characterization}: 
\code{\link{getCaseBinaryFeatures}()},
\code{\link{getCharacterizationDemographics}()},
\code{\link{getCohortDefinitions}()},
\code{\link{getDechallengeRechallenge}()},
\code{\link{getIncidenceRates}()},
\code{\link{getRiskFactors}()},
\code{\link{getTargetBinaryFeatures}()},
\code{\link{getTargetCounts}()},
\code{\link{getTimeToEvent}()},
\code{\link{plotAgeDistributions}()},
\code{\link{plotSexDistributions}()}
}
\concept{{Characterization}}
