% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimationQueries.R
\name{getCmMetaEstimation}
\alias{getCmMetaEstimation}
\title{Extract the cohort method meta analysis results}
\usage{
getCmMetaEstimation(
  connectionHandler,
  schema,
  cmTablePrefix = "cm_",
  cgTablePrefix = "cg_",
  esTablePrefix = "es_",
  targetIds = NULL,
  outcomeIds = NULL,
  comparatorIds = NULL
)
}
\arguments{
\item{connectionHandler}{A connection handler that connects to the database and extracts sql queries.  Create a connection handler via `ResultModelManager::ConnectionHandler$new()`.}

\item{schema}{The result database schema (e.g., 'main' for sqlite)}

\item{cmTablePrefix}{The prefix used for the cohort method results tables}

\item{cgTablePrefix}{The prefix used for the cohort generator results tables}

\item{esTablePrefix}{The prefix used for the evidence synthesis results tables}

\item{targetIds}{A vector of integers corresponding to the target cohort IDs}

\item{outcomeIds}{A vector of integers corresponding to the outcome cohort IDs}

\item{comparatorIds}{A vector of integers corresponding to the comparator cohort IDs}
}
\value{
Returns a data.frame with the columns:
\itemize{
 \item{databaseName} \description{the name of the database}
 \item{analysisId} \description{the analysis unique identifier}
 \item{description} \description{a description of the analysis}
 \item{targetName} \description{the target cohort name}
 \item{targetId} \description{the target cohort unique identifier}
 \item{comparatorName} \description{the comparator cohort name}
 \item{comparatorId} \description{the comparator cohort unique identifier}
 \item{outcomeName} \description{the outcome name}
 \item{outcomeId} \description{the outcome cohort unique identifier}
 
 \item{calibratedRr} \description{the calibrated relative risk}
 \item{calibratedRrCi95Lb} \description{the calibrated relative risk 95 percent confidence interval lower bound}
 \item{calibratedRrCi95Ub} \description{the calibrated relative risk 95 percent confidence interval upper bound}
 \item{calibratedP} \description{the two sided calibrated p value}
 \item{calibratedOneSidedP} \description{the one sided calibrated p value}
 \item{calibratedLogRr} \description{the calibrated relative risk logged}
 \item{calibratedSeLogRr} \description{the standard error of the calibrated relative risk logged}
 
 \item{targetSubjects} \description{the number of people in the target cohort across included database}
 \item{comparatorSubjects} \description{the number of people in the comparator cohort across included database}
 \item{targetDays} \description{the total number of days at risk across the target cohort people across included database}
 \item{comparatorDays} \description{the total number of days at risk across the comparator cohort people across included database}
 \item{targetOutcomes} \description{the total number of outcomes occuring during the time at risk for the target cohort people across included database}
 \item{comparatorOutcomes} \description{the total number of outcomes occuring during the time at risk for the comparator cohort people across included database}
 
 \item{nDatabases} \description{the number of databases included}
 }
}
\description{
This function extracts any meta analysis estimation results for cohort method.
}
\details{
Specify the connectionHandler, the schema and the target/comparator/outcome cohort IDs
}
\seealso{
Other {Estimation}: 
\code{\link{getCMEstimation}()},
\code{\link{getCmDiagnosticsData}()},
\code{\link{getSccsDiagnosticsData}()},
\code{\link{getSccsEstimation}()},
\code{\link{getSccsMetaEstimation}()}
}
\concept{{Estimation}}
