% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimationQueries.R
\name{getCmPropensityModel}
\alias{getCmPropensityModel}
\title{Extract the cohort method model}
\usage{
getCmPropensityModel(
  connectionHandler,
  schema,
  cmTablePrefix = "cm_",
  targetId = NULL,
  comparatorId = NULL,
  indicationId = NULL,
  analysisId = NULL,
  databaseId = NULL
)
}
\arguments{
\item{connectionHandler}{A connection handler that connects to the database and extracts sql queries.  Create a connection handler via `ResultModelManager::ConnectionHandler$new()`.}

\item{schema}{The result database schema (e.g., 'main' for sqlite)}

\item{cmTablePrefix}{The prefix used for the cohort method results tables}

\item{targetId}{An integer corresponding to the target cohort ID}

\item{comparatorId}{the comparator ID of interest}

\item{indicationId}{The indications that the target & comparator was nested to}

\item{analysisId}{the analysis ID to restrict to}

\item{databaseId}{the database ID to restrict to}
}
\value{
Returns a data.frame with the cohort method model
}
\description{
This function extracts the cohort method model.
}
\details{
Specify the connectionHandler, the schema and optionally the target/comparator/analysis/database IDs
}
\examples{
conDet <- getExampleConnectionDetails()

connectionHandler <- ResultModelManager::ConnectionHandler$new(conDet)

cmModel <- getCmPropensityModel(
  connectionHandler = connectionHandler, 
  schema = 'main'
)

}
\seealso{
Other Estimation:
\code{\link[=.getCmVersion]{.getCmVersion()}},
\code{\link[=.getSccsVersion]{.getSccsVersion()}},
\code{\link[=getCMEstimation]{getCMEstimation()}},
\code{\link[=getCmDiagnosticsData]{getCmDiagnosticsData()}},
\code{\link[=getCmMetaEstimation]{getCmMetaEstimation()}},
\code{\link[=getCmNegativeControlEstimates]{getCmNegativeControlEstimates()}},
\code{\link[=getCmOutcomes]{getCmOutcomes()}},
\code{\link[=getCmTable]{getCmTable()}},
\code{\link[=getCmTargets]{getCmTargets()}},
\code{\link[=getSccsDiagnosticsData]{getSccsDiagnosticsData()}},
\code{\link[=getSccsEstimation]{getSccsEstimation()}},
\code{\link[=getSccsMetaEstimation]{getSccsMetaEstimation()}},
\code{\link[=getSccsModel]{getSccsModel()}},
\code{\link[=getSccsNegativeControlEstimates]{getSccsNegativeControlEstimates()}},
\code{\link[=getSccsOutcomes]{getSccsOutcomes()}},
\code{\link[=getSccsTable]{getSccsTable()}},
\code{\link[=getSccsTargets]{getSccsTargets()}},
\code{\link[=getSccsTimeToEvent]{getSccsTimeToEvent()}},
\code{\link[=plotCmEstimates]{plotCmEstimates()}},
\code{\link[=plotSccsEstimates]{plotSccsEstimates()}}
}
\concept{Estimation}
