% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimationQueries.R
\name{getCmTable}
\alias{getCmTable}
\title{Extract the cohort method table specified}
\usage{
getCmTable(
  connectionHandler,
  schema,
  table = c("attrition", "follow_up_dist", "interaction_result", "covariate_balance",
    "kaplan_meier_dist", "likelihood_profile", "preference_score_dist",
    "propensity_model", "shared_covariate_balance")[1],
  cmTablePrefix = "cm_",
  cgTablePrefix = "cg_",
  databaseTable = "database_meta_data",
  targetIds = NULL,
  indicationIds = NULL,
  outcomeIds = NULL,
  comparatorIds = NULL,
  analysisIds = NULL,
  databaseIds = NULL
)
}
\arguments{
\item{connectionHandler}{A connection handler that connects to the database and extracts sql queries.  Create a connection handler via `ResultModelManager::ConnectionHandler$new()`.}

\item{schema}{The result database schema (e.g., 'main' for sqlite)}

\item{table}{The result table to extract}

\item{cmTablePrefix}{The prefix used for the cohort method results tables}

\item{cgTablePrefix}{The prefix used for the cohort generator results tables}

\item{databaseTable}{The name of the table with the database details (default 'database_meta_data')}

\item{targetIds}{A vector of integers corresponding to the target cohort IDs}

\item{indicationIds}{The indications that the target & comparator was nested to}

\item{outcomeIds}{A vector of integers corresponding to the outcome cohort IDs}

\item{comparatorIds}{A vector of integers corresponding to the comparator cohort IDs}

\item{analysisIds}{the analysis IDs to restrict to}

\item{databaseIds}{the database IDs to restrict to}
}
\value{
Returns a data.frame with the cohort method requested table
}
\description{
This function extracts the specific cohort method table.
}
\details{
Specify the connectionHandler, the schema and optionally the target/comparator/outcome/analysis/database IDs
}
\examples{
conDet <- getExampleConnectionDetails()

connectionHandler <- ResultModelManager::ConnectionHandler$new(conDet)

cmTable <- getCmTable(
  connectionHandler = connectionHandler, 
  schema = 'main',
  table = 'attrition'
)

}
\seealso{
Other Estimation:
\code{\link[=.getCmVersion]{.getCmVersion()}},
\code{\link[=.getSccsVersion]{.getSccsVersion()}},
\code{\link[=getCMEstimation]{getCMEstimation()}},
\code{\link[=getCmDiagnosticsData]{getCmDiagnosticsData()}},
\code{\link[=getCmMetaEstimation]{getCmMetaEstimation()}},
\code{\link[=getCmNegativeControlEstimates]{getCmNegativeControlEstimates()}},
\code{\link[=getCmOutcomes]{getCmOutcomes()}},
\code{\link[=getCmPropensityModel]{getCmPropensityModel()}},
\code{\link[=getCmTargets]{getCmTargets()}},
\code{\link[=getSccsDiagnosticsData]{getSccsDiagnosticsData()}},
\code{\link[=getSccsEstimation]{getSccsEstimation()}},
\code{\link[=getSccsMetaEstimation]{getSccsMetaEstimation()}},
\code{\link[=getSccsModel]{getSccsModel()}},
\code{\link[=getSccsNegativeControlEstimates]{getSccsNegativeControlEstimates()}},
\code{\link[=getSccsOutcomes]{getSccsOutcomes()}},
\code{\link[=getSccsTable]{getSccsTable()}},
\code{\link[=getSccsTargets]{getSccsTargets()}},
\code{\link[=getSccsTimeToEvent]{getSccsTimeToEvent()}},
\code{\link[=plotCmEstimates]{plotCmEstimates()}},
\code{\link[=plotSccsEstimates]{plotSccsEstimates()}}
}
\concept{Estimation}
