% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortQueries.R
\name{getCohortAttrition}
\alias{getCohortAttrition}
\title{Get cohort attrition}
\usage{
getCohortAttrition(
  connectionHandler,
  schema,
  cgTablePrefix = "cg_",
  databaseTable = "database_meta_data",
  cohortIds = NULL
)
}
\arguments{
\item{connectionHandler}{A connection handler object.}

\item{schema}{The database schema name.}

\item{cgTablePrefix}{Prefix for cohort generator tables. Default is 'cg_'.}

\item{databaseTable}{Name of the database metadata table. Default is 'database_meta_data'.}

\item{cohortIds}{Optional vector of cohort IDs to filter.}
}
\value{
A data.frame with attrition details for each cohort.
}
\description{
Retrieves attrition information for specified cohorts from the database.
}
\seealso{
Other Cohorts: 
\code{\link{getCohortCounts}()},
\code{\link{getCohortDefinitions}()},
\code{\link{getCohortInclusionRules}()},
\code{\link{getCohortInclusionStats}()},
\code{\link{getCohortInclusionSummary}()},
\code{\link{getCohortMeta}()},
\code{\link{getCohortSubsetAttrition}()},
\code{\link{getCohortSubsetDefinitions}()},
\code{\link{getSubsetText}()},
\code{\link{processCohortDefinitionsForQuarto}()},
\code{\link{processCohorts}()},
\code{\link{restrictCohortDefinitionsForQuarto}()}
}
\concept{Cohorts}
