% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortQueries.R
\name{getCohortAttrition}
\alias{getCohortAttrition}
\title{Get cohort attrition}
\usage{
getCohortAttrition(
  connectionHandler,
  schema,
  cgTablePrefix = "cg_",
  databaseTable = "database_meta_data",
  cohortIds = NULL
)
}
\arguments{
\item{connectionHandler}{A connection handler object.}

\item{schema}{The database schema name.}

\item{cgTablePrefix}{Prefix for cohort generator tables. Default is 'cg_'.}

\item{databaseTable}{Name of the database metadata table. Default is 'database_meta_data'.}

\item{cohortIds}{Optional vector of cohort IDs to filter.}
}
\value{
A tibble with attrition details for each cohort.
}
\description{
Retrieves attrition information for specified cohorts from the database.
}
