% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortQueries.R
\name{getCohortInclusionStats}
\alias{getCohortInclusionStats}
\title{Extract the cohort inclusion stats}
\usage{
getCohortInclusionStats(
  connectionHandler,
  schema,
  cgTablePrefix = "cg_",
  databaseTable = "database_meta_data",
  cohortIds = NULL
)
}
\arguments{
\item{connectionHandler}{A connection handler that connects to the database and extracts sql queries.  Create a connection handler via `ResultModelManager::ConnectionHandler$new()`.}

\item{schema}{The result database schema (e.g., 'main' for sqlite)}

\item{cgTablePrefix}{The prefix used for the cohort generator results tables}

\item{databaseTable}{The name of the table with the database details (default 'database_meta_data')}

\item{cohortIds}{Optionally a list of cohortIds to restrict to}
}
\value{
Returns a data.frame with the cohort inclusion stats
}
\description{
This function extracts all cohort inclusion stats for the cohorts of interest.
}
\details{
Specify the connectionHandler, the schema and the cohort IDs
}
\examples{
conDet <- getExampleConnectionDetails()

connectionHandler <- ResultModelManager::ConnectionHandler$new(conDet)

cohortInclsuionsStats <- getCohortInclusionStats(
  connectionHandler = connectionHandler, 
  schema = 'main'
)

}
\seealso{
Other Cohorts:
\code{\link[=getCohortAttrition]{getCohortAttrition()}},
\code{\link[=getCohortCounts]{getCohortCounts()}},
\code{\link[=getCohortDefinitions]{getCohortDefinitions()}},
\code{\link[=getCohortInclusionRules]{getCohortInclusionRules()}},
\code{\link[=getCohortInclusionSummary]{getCohortInclusionSummary()}},
\code{\link[=getCohortMeta]{getCohortMeta()}},
\code{\link[=getCohortSubsetAttrition]{getCohortSubsetAttrition()}},
\code{\link[=getCohortSubsetDefinitions]{getCohortSubsetDefinitions()}},
\code{\link[=getSubsetText]{getSubsetText()}},
\code{\link[=processCohortDefinitionsForQuarto]{processCohortDefinitionsForQuarto()}},
\code{\link[=processCohorts]{processCohorts()}},
\code{\link[=restrictCohortDefinitionsForQuarto]{restrictCohortDefinitionsForQuarto()}}
}
\concept{Cohorts}
