% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharacterzationQueries.R
\name{getContinuousTargetBaseline}
\alias{getContinuousTargetBaseline}
\title{Extract aggregate statistics of continuous feature analysis IDs of interest for targets}
\usage{
getContinuousTargetBaseline(
  connectionHandler,
  schema,
  cTablePrefix = "c_",
  cgTablePrefix = "cg_",
  databaseTable = "database_meta_data",
  characterizationTargetIds = NULL,
  analysisIds = NULL,
  databaseIds = NULL,
  includeNames = TRUE,
  minThreshold = NULL
)
}
\arguments{
\item{connectionHandler}{A connection handler that connects to the database and extracts sql queries.  Create a connection handler via `ResultModelManager::ConnectionHandler$new()`.}

\item{schema}{The result database schema (e.g., 'main' for sqlite)}

\item{cTablePrefix}{The prefix used for the characterization results tables}

\item{cgTablePrefix}{The prefix used for the cohort generator results tables}

\item{databaseTable}{The name of the table with the database details (default 'database_meta_data')}

\item{characterizationTargetIds}{The characterization target ids}

\item{analysisIds}{The feature extraction analysis ID of interest (e.g., 201 is condition)}

\item{databaseIds}{(Optional) A vector of database IDs to restrict to}

\item{includeNames}{Whether to add database and cohort names (setting to FALSE will make extraction quicker)}

\item{minThreshold}{(optional) The minimum average value for results to be returned}
}
\value{
Returns a data.frame with the columns:
\itemize{
 \item{databaseName the name of the database}
 \item{databaseId the unique identifier of the database}
 \item{targetName the target cohort name}
 \item{targetId the target cohort unique identifier}
 \item{minPriorObservation the minimum required observation days prior to index for an entry}
 \item{covariateName the name of the feature}
 \item{covariateId the id of the feature}
 \item{countValue the number of cases who have the feature}
 \item{minValue the minimum value observed for the feature}
 \item{maxValue the maximum value observed for the feature}
 \item{averageValue the mean value observed for the feature}
 \item{standardDeviation the standard deviation of the value observed for the feature}
 \item{medianValue the median value observed for the feature}
 \item{p10Value the 10th percentile of the value observed for the feature}
 \item{p25Value the 25th percentile of the value observed for the feature}
 \item{p75Value the 75th percentile of the value observed for the feature}
 \item{p90Value the 90th percentile of the value observed for the feature}
 
}
}
\description{
This function extracts the continuous feature extraction results for targets corresponding to specified target cohorts.
}
\details{
Specify the connectionHandler, the schema and the target/outcome cohort IDs
}
\examples{
conDet <- getExampleConnectionDetails()

connectionHandler <- ResultModelManager::ConnectionHandler$new(conDet)

tcf <- getContinuousTargetBaseline(
connectionHandler = connectionHandler, 
schema = 'main'
)

}
\seealso{
Other Characterization:
\code{\link[=characterizationCompareBinary]{characterizationCompareBinary()}},
\code{\link[=characterizationCompareContinuous]{characterizationCompareContinuous()}},
\code{\link[=getAggregateBinaryRiskFactors]{getAggregateBinaryRiskFactors()}},
\code{\link[=getAggregateContinuousRiskFactors]{getAggregateContinuousRiskFactors()}},
\code{\link[=getBinaryCaseSeries]{getBinaryCaseSeries()}},
\code{\link[=getBinaryRiskFactors]{getBinaryRiskFactors()}},
\code{\link[=getBinaryTargetBaseline]{getBinaryTargetBaseline()}},
\code{\link[=getCaseCounts]{getCaseCounts()}},
\code{\link[=getCharacterizationCaseSettings]{getCharacterizationCaseSettings()}},
\code{\link[=getCharacterizationDemographics]{getCharacterizationDemographics()}},
\code{\link[=getCharacterizationTargetSettings]{getCharacterizationTargetSettings()}},
\code{\link[=getContinuousCaseSeries]{getContinuousCaseSeries()}},
\code{\link[=getContinuousRiskFactors]{getContinuousRiskFactors()}},
\code{\link[=getDechallengeRechallenge]{getDechallengeRechallenge()}},
\code{\link[=getDechallengeRechallengeFails]{getDechallengeRechallengeFails()}},
\code{\link[=getIncidenceRates]{getIncidenceRates()}},
\code{\link[=getIncidenceTargetSettings]{getIncidenceTargetSettings()}},
\code{\link[=getNonCaseCounts]{getNonCaseCounts()}},
\code{\link[=getOutcomesUsedInCharacterization]{getOutcomesUsedInCharacterization()}},
\code{\link[=getOutcomesUsedInIncidence]{getOutcomesUsedInIncidence()}},
\code{\link[=getTargetCounts]{getTargetCounts()}},
\code{\link[=getTargetsUsedInCharacterization]{getTargetsUsedInCharacterization()}},
\code{\link[=getTargetsUsedInIncidence]{getTargetsUsedInIncidence()}},
\code{\link[=getTimeToEvent]{getTimeToEvent()}},
\code{\link[=plotAgeDistributions]{plotAgeDistributions()}},
\code{\link[=plotSexDistributions]{plotSexDistributions()}},
\code{\link[=viewIncidenceRate]{viewIncidenceRate()}}
}
\concept{Characterization}
