% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharacterzationQueries.R
\name{getIncidenceRates}
\alias{getIncidenceRates}
\title{Extract the cohort incidence result}
\usage{
getIncidenceRates(
  connectionHandler,
  schema,
  ciTablePrefix = "ci_",
  cgTablePrefix = "cg_",
  databaseTable = "database_meta_data",
  targetIds = NULL,
  outcomeIds = NULL
)
}
\arguments{
\item{connectionHandler}{A connection handler that connects to the database and extracts sql queries.  Create a connection handler via `ResultModelManager::ConnectionHandler$new()`.}

\item{schema}{The result database schema (e.g., 'main' for sqlite)}

\item{ciTablePrefix}{The prefix used for the cohort incidence results tables}

\item{cgTablePrefix}{The prefix used for the cohort generator results tables}

\item{databaseTable}{The name of the table with the database details (default 'database_meta_data')}

\item{targetIds}{A vector of integers corresponding to the target cohort IDs}

\item{outcomeIds}{A vector of integers corresponding to the outcome cohort IDs}
}
\value{
Returns a data.frame with the columns:
\itemize{
 \item{databaseName} \description{the name of the database}
 \item{targetName} \description{the target cohort name}
 \item{targetId} \description{the target cohort unique identifier}
 \item{outcomeName} \description{the outcome name}
 \item{outcomeId} \description{the outcome unique identifier}
 \item{cleanWindow} \description{clean windown around outcome}
 \item{subgroupName} \description{name for the result subgroup}
 \item{ageGroupName} \description{name for the result age group}
 \item{genderName} \description{name for the result gender group}
 \item{startYear} \description{name for the result start year}
 \item{tarStartWith} \description{time at risk start reference}
 \item{tarStartOffset} \description{time at risk start offset from reference}
 \item{tarEndWith} \description{time at risk end reference}
 \item{tarEndOffset} \description{time at risk end offset from reference}
 \item{personsAtRiskPe} \description{persons at risk per event}
 \item{personsAtRisk} \description{persons at risk}
 \item{personDaysPe} \description{person days per event}
 \item{personDays} \description{person days}
 \item{personOutcomesPe} \description{person outcome per event}
 \item{personOutcomes} \description{persons outcome}
 \item{outcomesPe} \description{number of outcome per event}
 \item{outcomes} \description{number of outcome}
 \item{incidenceProportionP100p} \description{incidence proportion per 100 persons}
 \item{incidenceRateP100py} \description{incidence rate per 100 person years}
 }
}
\description{
This function extracts all incidence rates across databases in the results for specified target and outcome cohorts.
}
\details{
Specify the connectionHandler, the schema and the target/outcome cohort IDs
}
\seealso{
Other {Characterization}: 
\code{\link{getCaseBinaryFeatures}()},
\code{\link{getCaseCounts}()},
\code{\link{getCohortDefinitions}()},
\code{\link{getDechallengeRechallenge}()},
\code{\link{getRiskFactors}()},
\code{\link{getTargetBinaryFeatures}()},
\code{\link{getTargetCounts}()},
\code{\link{getTimeToEvent}()}
}
\concept{{Characterization}}
