% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharacterzationQueries.R
\name{getNonCaseCounts}
\alias{getNonCaseCounts}
\title{Extract the non-cases counts result}
\usage{
getNonCaseCounts(
  connectionHandler,
  schema,
  cTablePrefix = "c_",
  cgTablePrefix = "cg_",
  databaseTable = "database_meta_data",
  characterizationTargetIds = NULL,
  characterizationCaseIds = NULL,
  startAnchor = NULL,
  endAnchor = NULL,
  riskWindowStart = NULL,
  riskWindowEnd = NULL,
  outcomeIds = NULL,
  outcomeWashouts = NULL,
  databaseIds = NULL,
  includeNames = TRUE
)
}
\arguments{
\item{connectionHandler}{A connection handler that connects to the database and extracts sql queries.  Create a connection handler via `ResultModelManager::ConnectionHandler$new()`.}

\item{schema}{The result database schema (e.g., 'main' for sqlite)}

\item{cTablePrefix}{The prefix used for the characterization results tables}

\item{cgTablePrefix}{The prefix used for the cohort generator results tables}

\item{databaseTable}{The name of the table with the database details (default 'database_meta_data')}

\item{characterizationTargetIds}{The characterization target cohort ids of interest}

\item{characterizationCaseIds}{The characterization case ids of interest}

\item{startAnchor}{optional filter by startAchor}

\item{endAnchor}{optional filter by endAchor}

\item{riskWindowStart}{optional filter by riskWindowStart}

\item{riskWindowEnd}{optional filter by riskWindowEnd}

\item{outcomeIds}{A vector of integers corresponding to the outcome cohort IDs}

\item{outcomeWashouts}{optional filter by washout}

\item{databaseIds}{A vector of database IDs to restrict to}

\item{includeNames}{whether to add the database names and cohort names}
}
\value{
Returns a data.frame with the columns:
\itemize{
 \item{databaseName the name of the database}
 \item{databaseId the unique identifier of the database}
 \item{characterizationCaseId the unique identifier of target, outcome and TAR combination}
 \item{targetName the target cohort name}
 \item{targetId the target cohort unique identifier}
 \item{limitToFirstInNDays target index is limited to first in N days}
 \item{minPriorObservation the minimum required observation days prior to index for an entry}
 \item{nestingCohortId the cohort id a person must be in at index}
 \item{nestingName the cohort name a person must be in at index}
 \item{minAge min age to be included at index}
 \item{maxAge max age to be included at index}
 \item{studyStart index must be on or after this date to be included}
 \item{studyEnd index must be on or before this date to be included}
 \item{genderConceptIds the gender concept ids a subject must have to be included}
 \item{outcomeName the outcome name}
 \item{outcomeId the outcome unique identifier}
 \item{outcomeWashoutDays patients with the outcome occurring within this number of days prior to index are excluded (NA means no exclusion)}
 \item{rowCount the number of entries in the cohort}
 \item{personCount the number of people in the cohort}
 \item{withoutExcludedPersonCount the number of people in the target ignoring exclusions}
 }
}
\description{
This function extracts non-case counts across databases in the results for specified target and outcome cohorts.
}
\details{
Specify the connectionHandler, the schema and the target/outcome cohort IDs
}
\examples{
conDet <- getExampleConnectionDetails()

connectionHandler <- ResultModelManager::ConnectionHandler$new(conDet)

tc <- getNonCaseCounts(
connectionHandler = connectionHandler, 
schema = 'main'
)

}
\seealso{
Other Characterization:
\code{\link[=characterizationCompareBinary]{characterizationCompareBinary()}},
\code{\link[=characterizationCompareContinuous]{characterizationCompareContinuous()}},
\code{\link[=getAggregateBinaryRiskFactors]{getAggregateBinaryRiskFactors()}},
\code{\link[=getAggregateContinuousRiskFactors]{getAggregateContinuousRiskFactors()}},
\code{\link[=getBinaryCaseSeries]{getBinaryCaseSeries()}},
\code{\link[=getBinaryRiskFactors]{getBinaryRiskFactors()}},
\code{\link[=getBinaryTargetBaseline]{getBinaryTargetBaseline()}},
\code{\link[=getCaseCounts]{getCaseCounts()}},
\code{\link[=getCharacterizationCaseSettings]{getCharacterizationCaseSettings()}},
\code{\link[=getCharacterizationDemographics]{getCharacterizationDemographics()}},
\code{\link[=getCharacterizationTargetSettings]{getCharacterizationTargetSettings()}},
\code{\link[=getContinuousCaseSeries]{getContinuousCaseSeries()}},
\code{\link[=getContinuousRiskFactors]{getContinuousRiskFactors()}},
\code{\link[=getContinuousTargetBaseline]{getContinuousTargetBaseline()}},
\code{\link[=getDechallengeRechallenge]{getDechallengeRechallenge()}},
\code{\link[=getDechallengeRechallengeFails]{getDechallengeRechallengeFails()}},
\code{\link[=getIncidenceRates]{getIncidenceRates()}},
\code{\link[=getIncidenceTargetSettings]{getIncidenceTargetSettings()}},
\code{\link[=getOutcomesUsedInCharacterization]{getOutcomesUsedInCharacterization()}},
\code{\link[=getOutcomesUsedInIncidence]{getOutcomesUsedInIncidence()}},
\code{\link[=getTargetCounts]{getTargetCounts()}},
\code{\link[=getTargetsUsedInCharacterization]{getTargetsUsedInCharacterization()}},
\code{\link[=getTargetsUsedInIncidence]{getTargetsUsedInIncidence()}},
\code{\link[=getTimeToEvent]{getTimeToEvent()}},
\code{\link[=plotAgeDistributions]{plotAgeDistributions()}},
\code{\link[=plotSexDistributions]{plotSexDistributions()}},
\code{\link[=viewIncidenceRate]{viewIncidenceRate()}}
}
\concept{Characterization}
