% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharacterzationQueries.R
\name{getOutcomesUsedInCharacterization}
\alias{getOutcomesUsedInCharacterization}
\title{A function to extract the outcomes found in characterization}
\usage{
getOutcomesUsedInCharacterization(
  connectionHandler,
  schema,
  cTablePrefix = "c_",
  cgTablePrefix = "cg_",
  characterizationTargetId = NULL,
  targetId = NULL,
  printTimes = FALSE,
  useDcrc = TRUE,
  useTte = TRUE,
  useRf = TRUE,
  useCs = TRUE
)
}
\arguments{
\item{connectionHandler}{A connection handler that connects to the database and extracts sql queries.  Create a connection handler via `ResultModelManager::ConnectionHandler$new()`.}

\item{schema}{The result database schema (e.g., 'main' for sqlite)}

\item{cTablePrefix}{The prefix used for the characterization results tables}

\item{cgTablePrefix}{The prefix used for the cohort generator results tables}

\item{characterizationTargetId}{A vector of characterizationTargetIds to restrict to}

\item{targetId}{An integer corresponding to the target cohort ID}

\item{printTimes}{Print the time it takes to run each query}

\item{useDcrc}{look for outcome in dechal-rechal results}

\item{useTte}{look for outcome in time-to-event results}

\item{useRf}{look for outcome in risk-factor results}

\item{useCs}{look for outcome in case series results}
}
\value{
A data.frame with the characterization outcome cohort ids, names and which characterization analyses the cohorts are used in.
}
\description{
A function to extract the outcomes found in characterization
}
\details{
Specify the connectionHandler, the schema and the prefixes
}
\examples{
conDet <- getExampleConnectionDetails()

connectionHandler <- ResultModelManager::ConnectionHandler$new(conDet)

cohorts <- getOutcomesUsedInCharacterization(
  connectionHandler = connectionHandler, 
  schema = 'main'
)

}
\seealso{
Other Characterization: 
\code{\link{characterizationCompareBinary}()},
\code{\link{characterizationCompareContinuous}()},
\code{\link{getBinaryCaseSeries}()},
\code{\link{getBinaryRiskFactors}()},
\code{\link{getBinaryTargetBaseline}()},
\code{\link{getCaseCounts}()},
\code{\link{getCharacterizationCaseSettings}()},
\code{\link{getCharacterizationDemographics}()},
\code{\link{getCharacterizationTargetSettings}()},
\code{\link{getContinuousCaseSeries}()},
\code{\link{getContinuousRiskFactors}()},
\code{\link{getContinuousTargetBaseline}()},
\code{\link{getDechallengeRechallenge}()},
\code{\link{getDechallengeRechallengeFails}()},
\code{\link{getIncidenceRates}()},
\code{\link{getIncidenceTargetSettings}()},
\code{\link{getNonCaseCounts}()},
\code{\link{getOutcomesUsedInIncidence}()},
\code{\link{getTargetCounts}()},
\code{\link{getTargetsUsedInCharacterization}()},
\code{\link{getTargetsUsedInIncidence}()},
\code{\link{getTimeToEvent}()},
\code{\link{plotAgeDistributions}()},
\code{\link{plotSexDistributions}()},
\code{\link{viewIncidenceRate}()}
}
\concept{Characterization}
