% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharacterzationQueries.R
\name{getOutcomesUsedInIncidence}
\alias{getOutcomesUsedInIncidence}
\title{A function to extract the outcomes found in incidence}
\usage{
getOutcomesUsedInIncidence(
  connectionHandler,
  schema,
  ciTablePrefix = "ci_",
  cgTablePrefix = "cg_",
  targetId = NULL,
  parentId = NULL
)
}
\arguments{
\item{connectionHandler}{A connection handler that connects to the database and extracts sql queries.  Create a connection handler via `ResultModelManager::ConnectionHandler$new()`.}

\item{schema}{The result database schema (e.g., 'main' for sqlite)}

\item{ciTablePrefix}{The prefix used for the cohort incidence results tables}

\item{cgTablePrefix}{The prefix used for the cohort generator results tables}

\item{targetId}{An integer corresponding to the target cohort ID}

\item{parentId}{the parent target cohort Id to extract outcomes for}
}
\value{
A data.frame with the incidence outcome cohort ids and names
}
\description{
A function to extract the outcomes found in incidence
}
\details{
Specify the connectionHandler, the schema and the prefixes
}
\examples{
conDet <- getExampleConnectionDetails()

connectionHandler <- ResultModelManager::ConnectionHandler$new(conDet)

outcomes <- getOutcomesUsedInIncidence(
  connectionHandler = connectionHandler, 
  schema = 'main'
)

}
\seealso{
Other Characterization: 
\code{\link{characterizationCompareBinary}()},
\code{\link{characterizationCompareContinuous}()},
\code{\link{getBinaryCaseSeries}()},
\code{\link{getBinaryRiskFactors}()},
\code{\link{getBinaryTargetBaseline}()},
\code{\link{getCaseCounts}()},
\code{\link{getCharacterizationCaseSettings}()},
\code{\link{getCharacterizationDemographics}()},
\code{\link{getCharacterizationTargetSettings}()},
\code{\link{getContinuousCaseSeries}()},
\code{\link{getContinuousRiskFactors}()},
\code{\link{getContinuousTargetBaseline}()},
\code{\link{getDechallengeRechallenge}()},
\code{\link{getDechallengeRechallengeFails}()},
\code{\link{getIncidenceRates}()},
\code{\link{getIncidenceTargetSettings}()},
\code{\link{getNonCaseCounts}()},
\code{\link{getOutcomesUsedInCharacterization}()},
\code{\link{getTargetCounts}()},
\code{\link{getTargetsUsedInCharacterization}()},
\code{\link{getTargetsUsedInIncidence}()},
\code{\link{getTimeToEvent}()},
\code{\link{plotAgeDistributions}()},
\code{\link{plotSexDistributions}()},
\code{\link{viewIncidenceRate}()}
}
\concept{Characterization}
