% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionQueries.R
\name{getPredictionCohorts}
\alias{getPredictionCohorts}
\title{Extract a complete set of cohorts used in the prediction results}
\usage{
getPredictionCohorts(connectionHandler, resultDatabaseSettings)
}
\arguments{
\item{connectionHandler}{A connection handler that connects to the database and extracts sql queries.  Create a connection handler via `ResultModelManager::ConnectionHandler$new()`.}

\item{resultDatabaseSettings}{A list containing the schema and table prefixes created using `OhdsiShinyAppBuilder::createDefaultResultDatabaseSettings()`}
}
\value{
Returns a data.frame with the columns: 
\itemize{
 \item{cohortId} \description{the cohort definition ID}
 \item{cohortName} \description{the name of the cohort}
 \item{type} \description{whether the cohort was used as a target or outcome cohort}
}
}
\description{
This function extracts the target and outcome cohorts used to develop any model in the results
}
\details{
Specify the connectionHandler, the resultDatabaseSettings and any targetIds or outcomeIds to restrict models to
}
\seealso{
Other {Prediction}: 
\code{\link{getPredictionDiagnosticTable}()},
\code{\link{getPredictionDiagnostics}()},
\code{\link{getPredictionHyperParamSearch}()},
\code{\link{getPredictionIntercept}()},
\code{\link{getPredictionModelDesigns}()},
\code{\link{getPredictionPerformanceTable}()},
\code{\link{getPredictionPerformances}()},
\code{\link{getPredictionTopPredictors}()}
}
\concept{{Prediction}}
