% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionQueries.R
\name{getPredictionDiagnosticTable}
\alias{getPredictionDiagnosticTable}
\title{Extract specific diagnostic table}
\usage{
getPredictionDiagnosticTable(
  connectionHandler,
  resultDatabaseSettings,
  table = "diagnostic_participants",
  diagnosticId = NULL
)
}
\arguments{
\item{connectionHandler}{A connection handler that connects to the database and extracts sql queries.  Create a connection handler via `ResultModelManager::ConnectionHandler$new()`.}

\item{resultDatabaseSettings}{A list containing the schema and table prefixes created using `OhdsiShinyAppBuilder::createDefaultResultDatabaseSettings()`}

\item{table}{The table to extract}

\item{diagnosticId}{(optional) restrict to the input diagnosticId}
}
\value{
Returns a data.frame with the specified table
}
\description{
This function extracts the specified diagnostic table
}
\details{
Specify the connectionHandler, the resultDatabaseSettings, the table of interest and (optionally) a diagnosticId to filter to
}
\seealso{
Other {Prediction}: 
\code{\link{getPredictionCohorts}()},
\code{\link{getPredictionDiagnostics}()},
\code{\link{getPredictionHyperParamSearch}()},
\code{\link{getPredictionIntercept}()},
\code{\link{getPredictionModelDesigns}()},
\code{\link{getPredictionPerformanceTable}()},
\code{\link{getPredictionPerformances}()},
\code{\link{getPredictionTopPredictors}()}
}
\concept{{Prediction}}
