% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionQueries.R
\name{getPredictionModelDesigns}
\alias{getPredictionModelDesigns}
\title{Extract the model designs and aggregate performances for the prediction results}
\usage{
getPredictionModelDesigns(
  connectionHandler,
  schema,
  plpTablePrefix = "plp_",
  cgTablePrefix = "cg_",
  targetIds = NULL,
  outcomeIds = NULL
)
}
\arguments{
\item{connectionHandler}{A connection handler that connects to the database and extracts sql queries.  Create a connection handler via `ResultModelManager::ConnectionHandler$new()`.}

\item{schema}{The result database schema (e.g., 'main' for sqlite)}

\item{plpTablePrefix}{The prefix used for the patient level prediction results tables}

\item{cgTablePrefix}{The prefix used for the cohort generator results tables}

\item{targetIds}{A vector of integers corresponding to the target cohort IDs}

\item{outcomeIds}{A vector of integers corresponding to the outcome cohort IDs}
}
\value{
Returns a data.frame with the columns:
\itemize{ 
 \item{modelDesignId} \description{a unique identifier in the database for the model design}
 \item{modelType} \description{the type of classifier or surival model}
 \item{developmentTargetId} \description{a unique identifier for the development target ID}
 \item{developmentTargetName} \description{the name of the development target cohort}
 \item{developmentTargetJson} \description{the json of the target cohort}
 \item{developmentOutcomeId} \description{a unique identifier for the development outcome ID}
 \item{developmentOutcomeName} \description{the name of the development outcome cohort}
 \item{timeAtRisk} \description{the time at risk string}
 \item{developmentOutcomeJson} \description{the json of the outcome cohort}
 \item{covariateSettingsJson} \description{the covariate settings json}
 \item{populationSettingsJson} \description{the population settings json}
 \item{tidyCovariatesSettingsJson} \description{the tidy covariate settings json}
 \item{plpDataSettingsJson} \description{the plp data extraction settings json}
 \item{featureEngineeringSettingsJson} \description{the feature engineering settings json}
 \item{splitSettingsJson} \description{the split settings json}
 \item{sampleSettingsJson} \description{the sample settings json}
 \item{minAuroc} \description{the min AUROC value of models developed using the model design across databases}
 \item{meanAuroc} \description{the mean AUROC value of models developed using the model design across databases}
 \item{maxAuroc} \description{the max AUROC value of models developed using the model design across databases}
 \item{noDiagnosticDatabases} \description{the number of databases where the model design diagnostics were generated}
 \item{noDevelopmentDatabases} \description{the number of databases where the model design was used to develop models}
 \item{noValidationDatabases} \description{the number of databases where the models developed using the model design was externally validated}
}
}
\description{
This function extracts the model design settings and min/max/mean AUROC values of the models developed
using the model design across databases
}
\details{
Specify the connectionHandler, the resultDatabaseSettings and (optionally) any targetIds or outcomeIds to restrict model designs to
}
\seealso{
Other {Prediction}: 
\code{\link{getPredictionCohorts}()},
\code{\link{getPredictionDiagnosticTable}()},
\code{\link{getPredictionDiagnostics}()},
\code{\link{getPredictionHyperParamSearch}()},
\code{\link{getPredictionIntercept}()},
\code{\link{getPredictionPerformanceTable}()},
\code{\link{getPredictionPerformances}()},
\code{\link{getPredictionTopPredictors}()}
}
\concept{{Prediction}}
