% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionQueries.R
\name{getPredictionTopPredictors}
\alias{getPredictionTopPredictors}
\title{Extract the top N predictors per model}
\usage{
getPredictionTopPredictors(
  connectionHandler,
  resultDatabaseSettings,
  targetIds = NULL,
  outcomeIds = NULL,
  numberPredictors = 100
)
}
\arguments{
\item{connectionHandler}{A connection handler that connects to the database and extracts sql queries.  Create a connection handler via `ResultModelManager::ConnectionHandler$new()`.}

\item{resultDatabaseSettings}{A list containing the schema and table prefixes created using `OhdsiShinyAppBuilder::createDefaultResultDatabaseSettings()`}

\item{targetIds}{A vector of integers corresponding to the target cohort IDs}

\item{outcomeIds}{A vector of integers corresponding to the outcome cohort IDs}

\item{numberPredictors}{the number of predictors per model to return}
}
\value{
Returns a data.frame with the columns: 
\itemize{
 \item{databaseName} \description{the name of the database the model was developed on}
 \item{tarStartDay} \description{the time-at-risk start day}
 \item{tarStartAnchor} \description{whether the time-at-risk start is relative to cohort start or end}
 \item{tarEndDay} \description{the time-at-risk end day}
 \item{tarEndAnchor} \description{whether the time-at-risk end is relative to cohort start or end}
 \item{performanceId} \description{a unique identifier for the performance}
 \item{covariateId} \description{the FeatureExtraction covariate identifier}
 \item{covariateName} \description{the name of the covariate}
 \item{conceptId} \description{the covariates corresponding concept or 0}
 \item{covariateValue} \description{the feature importance or coefficient value}
 \item{covariateCount} \description{how many people had the covariate}
 \item{covariateMean} \description{the fraction of the target population with the covariate}
 \item{covariateStDev} \description{the standard deviation}
 \item{withNoOutcomeCovariateCount} \description{the number of the target population without the outcome with the covariate}
 \item{withNoOutcomeCovariateMean} \description{the fraction of the target population without the outcome with the covariate}
 \item{withNoOutcomeCovariateStDev} \description{the covariate standard deviation of the target population without the outcome}
 \item{withOutcomeCovariateCount} \description{the number of the target population with the outcome with the covariate}
 \item{withOutcomeCovariateMean} \description{the fraction of the target population with the outcome with the covariate}
 \item{withOutcomeCovariateStDev} \description{the covariate standard deviation of the target population with the outcome}
 \item{standardizedMeanDiff} \description{the standardized mean difference comparing the target population with outcome and without the outcome}
 \item{rn} \description{the row number showing the covariate rank}
}
}
\description{
This function extracts the top N predictors per model from the prediction results tables
}
\details{
Specify the connectionHandler, the resultDatabaseSettings and (optionally) any targetIds or outcomeIds to restrict models to
}
\seealso{
Other {Prediction}: 
\code{\link{getPredictionCohorts}()},
\code{\link{getPredictionDiagnosticTable}()},
\code{\link{getPredictionDiagnostics}()},
\code{\link{getPredictionHyperParamSearch}()},
\code{\link{getPredictionIntercept}()},
\code{\link{getPredictionModelDesigns}()},
\code{\link{getPredictionPerformanceTable}()},
\code{\link{getPredictionPerformances}()}
}
\concept{{Prediction}}
