% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SccQueries.R
\name{getSccMetaEstimation}
\alias{getSccMetaEstimation}
\title{Extract the self controlled cohort meta analytic (evidence synthesis) results}
\usage{
getSccMetaEstimation(
  connectionHandler,
  schema,
  sccTablePrefix = "scc_",
  cgTablePrefix = "cg_",
  esTablePrefix = "es_",
  analysisIds = NULL,
  targetIds = NULL,
  outcomeIds = NULL,
  evidenceSynthesisAnalysisIds = NULL
)
}
\arguments{
\item{connectionHandler}{A connection handler that connects to the database and extracts sql queries.  Create a connection handler via `ResultModelManager::ConnectionHandler$new()`.}

\item{schema}{The result database schema (e.g., 'main' for sqlite)}

\item{sccTablePrefix}{The prefix used for the self controlled cohort results tables}

\item{cgTablePrefix}{The prefix used for the cohort generator results tables}

\item{esTablePrefix}{The prefix used for the evidence synthesis results tables}

\item{analysisIds}{A vector of analysis ids to restrict the results to}

\item{targetIds}{A vector of target cohort ids to restrict the results to}

\item{outcomeIds}{A vector of outcome cohort ids to restrict the results to}

\item{evidenceSynthesisAnalysisIds}{A vector of evidence synthesis analysis ids to restrict the results to}
}
\value{

}
\description{
Returns the meta analytic (evidence synthesis) effect estimates for the
selected exposure (target) and outcome cohort pairs.  This function reads
the es_scc_* tables created by running the evidence synthesis module on
self controlled cohort results.
}
\details{
Specify the connectionHandler, the schema and the optional target/outcome/analysis
cohort identifiers to restrict the results
}
\seealso{
Other SelfControlledCohort: 
\code{\link{createSccSignalsTable}()},
\code{\link{getSccAnalysisSettings}()},
\code{\link{getSccDiagnosticsData}()},
\code{\link{getSccEstimation}()},
\code{\link{getSccMetaExploration}()},
\code{\link{getSccNegativeControlEstimates}()},
\code{\link{getSccOutcomes}()},
\code{\link{getSccSignals}()},
\code{\link{getSccSummaryStats}()},
\code{\link{getSccTargets}()},
\code{\link{plotSccBoxPlot}()},
\code{\link{plotSccForest}()},
\code{\link{plotSccSystematicError}()}
}
\concept{SelfControlledCohort}
