% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimationQueries.R
\name{getSccsDiagnosticsData}
\alias{getSccsDiagnosticsData}
\title{Extract the self controlled case series (sccs) diagostic results}
\usage{
getSccsDiagnosticsData(
  connectionHandler,
  schema,
  sccsTablePrefix = "sccs_",
  cgTablePrefix = "cg_",
  databaseTable = "database_meta_data",
  targetIds = NULL,
  outcomeIds = NULL
)
}
\arguments{
\item{connectionHandler}{A connection handler that connects to the database and extracts sql queries.  Create a connection handler via `ResultModelManager::ConnectionHandler$new()`.}

\item{schema}{The result database schema (e.g., 'main' for sqlite)}

\item{sccsTablePrefix}{The prefix used for the cohort generator results tables}

\item{cgTablePrefix}{The prefix used for the cohort generator results tables}

\item{databaseTable}{The name of the table with the database details (default 'database_meta_data')}

\item{targetIds}{A vector of integers corresponding to the target cohort IDs}

\item{outcomeIds}{A vector of integers corresponding to the outcome cohort IDs}
}
\value{
Returns a data.frame with the columns:
\itemize{
 \item{databaseName} \description{the database name}
 \item{analysisId} \description{the analysis unique identifier}
 \item{description} \description{an analysis description}
 \item{targetName} \description{the target name}
 \item{targetId} \description{the target cohort id}
 \item{outcomeName} \description{the outcome name}
 \item{outcomeId} \description{the outcome cohort id}
 \item{covariateName} \description{whether main or secondary analysis}
 \item{mdrr} \description{the maximum passable minimum detectable relative risk (mdrr) value.  If the mdrr is greater than this the diagnostics will fail.}
 \item{ease} \description{The expected absolute systematic error (ease) measures residual bias.}
 \item{timeTrendP} \description{The p for whether the mean monthly ratio between observed and expected is no greater than 1.25.}
 \item{preExposureP} \description{One-sided p-value for whether the rate before expore is higher than after, against the null of no difference.}
 \item{mdrrDiagnostic} \description{whether the mdrr (power) diagnostic passed or failed.}
 \item{easeDiagnostic} \description{whether the ease diagnostic passed or failed.}
 \item{timeTrendDiagnostic} \description{Pass / warning / fail / not evaluated classification of the time trend (unstalbe months) diagnostic.}
 \item{preExposureDiagnostic} \description{Pass / warning / fail / not evaluated classification of the time trend (unstalbe months) diagnostic.}
 \item{unblind} \description{whether the results can be unblinded.}
 \item{unblindForEvidenceSynthesis} \description{whether the results can be unblinded for the meta analysis.}
 \item{summaryValue} \description{summary of diagnostics results. FAIL, PASS or number of warnings.}
 }
}
\description{
This function extracts the sccs diagnostics that examine whether the analyses were sufficiently powered
and checks for different types of bias.
}
\details{
Specify the connectionHandler, the schema and the target/outcome cohort IDs
}
\seealso{
Other {Estimation}: 
\code{\link{getCMEstimation}()},
\code{\link{getCmDiagnosticsData}()},
\code{\link{getCmMetaEstimation}()},
\code{\link{getSccsEstimation}()},
\code{\link{getSccsMetaEstimation}()}
}
\concept{{Estimation}}
