% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimationQueries.R
\name{getSccsEstimation}
\alias{getSccsEstimation}
\title{Extract the self controlled case series (sccs) results}
\usage{
getSccsEstimation(
  connectionHandler,
  schema,
  sccsTablePrefix = "sccs_",
  cgTablePrefix = "cg_",
  databaseTable = "database_meta_data",
  targetIds = NULL,
  outcomeIds = NULL
)
}
\arguments{
\item{connectionHandler}{A connection handler that connects to the database and extracts sql queries.  Create a connection handler via `ResultModelManager::ConnectionHandler$new()`.}

\item{schema}{The result database schema (e.g., 'main' for sqlite)}

\item{sccsTablePrefix}{The prefix used for the cohort generator results tables}

\item{cgTablePrefix}{The prefix used for the cohort generator results tables}

\item{databaseTable}{The name of the table with the database details (default 'database_meta_data')}

\item{targetIds}{A vector of integers corresponding to the target cohort IDs}

\item{outcomeIds}{A vector of integers corresponding to the outcome cohort IDs}
}
\value{
Returns a data.frame with the columns:
\itemize{
 \item{databaseName} \description{the database name}
 \item{analysisId} \description{the analysis unique identifier}
 \item{description} \description{an analysis description}
 \item{targetName} \description{the target name}
 \item{targetId} \description{the target cohort id}
 \item{outcomeName} \description{the outcome name}
 \item{outcomeId} \description{the outcome cohort id}
 \item{covariateName} \description{whether main or secondary analysis}
 \item{outcomeSubjects} \description{The number of subjects with at least one outcome.}
 \item{outcomeEvents} \description{The number of outcome events.}
 \item{outcomeObservationPeriods} \description{The number of observation periods containing at least one outcome.}
 \item{covariateSubjects} \description{The number of subjects having the covariate.}
 \item{covariateDays} \description{The total covariate time in days.}
 \item{covariateEras} \description{The number of continuous eras of the covariate.}
 \item{covariateOutcomes} \description{The number of outcomes observed during the covariate time.}
 \item{observedDays} \description{The number of days subjects were observed.}
 \item{rr} \description{the relative risk}
 \item{ci95Lb} \description{the lower bound of the 95 percent confidence interval for the relative risk}
 \item{ci95Ub} \description{the upper bound of the 95 percent confidence interval for the relative risk}
 \item{p} \description{the p-value for the relative risk}
 \item{logRr} \description{the log of the relative risk}
 \item{seLogRr} \description{the standard error or the log of the relative risk}
 \item{calibratedRr} \description{the calibrated relative risk}
 \item{calibratedCi95Lb} \description{the lower bound of the 95 percent confidence interval for the calibrated relative risk}
 \item{calibratedCi95Ub} \description{the upper bound of the 95 percent confidence interval for the calibrated relative risk}
 \item{calibratedP} \description{the calibrated p-value}
 \item{calibratedLogRr} \description{the calibrated log of the relative risk}
 \item{calibratedSeLogRr} \description{the calibrated log of the relative risk standard error}
 \item{llr} \description{The log of the likelihood ratio (of the MLE vs the null hypothesis of no effect).}
 \item{mdrr} \description{The minimum detectable relative risk.}
 \item{unblind} \description{Whether the results can be unblinded}
 }
}
\description{
This function extracts the single database sccs estimates
}
\details{
Specify the connectionHandler, the schema and the target/outcome cohort IDs
}
\seealso{
Other {Estimation}: 
\code{\link{getCMEstimation}()},
\code{\link{getCmDiagnosticsData}()},
\code{\link{getCmMetaEstimation}()},
\code{\link{getSccsDiagnosticsData}()},
\code{\link{getSccsMetaEstimation}()},
\code{\link{plotCmEstimates}()},
\code{\link{plotSccsEstimates}()}
}
\concept{{Estimation}}
