% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimationQueries.R
\name{getSccsMetaEstimation}
\alias{getSccsMetaEstimation}
\title{Extract the self controlled case series (sccs) meta analysis results}
\usage{
getSccsMetaEstimation(
  connectionHandler,
  schema,
  sccsTablePrefix = "sccs_",
  cgTablePrefix = "cg_",
  esTablePrefix = "es_",
  targetIds = NULL,
  outcomeIds = NULL
)
}
\arguments{
\item{connectionHandler}{A connection handler that connects to the database and extracts sql queries.  Create a connection handler via `ResultModelManager::ConnectionHandler$new()`.}

\item{schema}{The result database schema (e.g., 'main' for sqlite)}

\item{sccsTablePrefix}{The prefix used for the cohort generator results tables}

\item{cgTablePrefix}{The prefix used for the cohort generator results tables}

\item{esTablePrefix}{The prefix used for the evidence synthesis results tables}

\item{targetIds}{A vector of integers corresponding to the target cohort IDs}

\item{outcomeIds}{A vector of integers corresponding to the outcome cohort IDs}
}
\value{
Returns a data.frame with the columns:
\itemize{
 #'  \item{databaseName} \description{the database name}
 \item{analysisId} \description{the analysis unique identifier}
 \item{description} \description{an analysis description}
 \item{targetName} \description{the target name}
 \item{targetId} \description{the target cohort id}
 \item{outcomeName} \description{the outcome name}
 \item{outcomeId} \description{the outcome cohort id}
 \item{covariateName} \description{whether main or secondary analysis}
 \item{outcomeSubjects} \description{The number of subjects with at least one outcome.}
 \item{outcomeEvents} \description{The number of outcome events.}
 \item{outcomeObservationPeriods} \description{The number of observation periods containing at least one outcome.}
 \item{covariateSubjects} \description{The number of subjects having the covariate.}
 \item{covariateDays} \description{The total covariate time in days.}
 \item{covariateEras} \description{The number of continuous eras of the covariate.}
 \item{covariateOutcomes} \description{The number of outcomes observed during the covariate time.}
 \item{observedDays} \description{The number of days subjects were observed.}
 \item{calibratedRr} \description{the calibrated relative risk}
 \item{calibratedCi95Lb} \description{the lower bound of the 95 percent confidence interval for the calibrated relative risk}
 \item{calibratedCi95Ub} \description{the upper bound of the 95 percent confidence interval for the calibrated relative risk}
 \item{calibratedP} \description{the calibrated p-value}
 \item{calibratedOneSidedP} \description{the calibrated one sided p-value}
 \item{calibratedLogRr} \description{the calibrated log of the relative risk}
 \item{calibratedSeLogRr} \description{the calibrated log of the relative risk standard error}
 \item{nDatabases} \description{The number of databases included in the estimate.}
 }
}
\description{
This function extracts any meta analysis estimation results for sccs.
}
\details{
Specify the connectionHandler, the schema and the targetoutcome cohort IDs
}
\seealso{
Other {Estimation}: 
\code{\link{getCMEstimation}()},
\code{\link{getCmDiagnosticsData}()},
\code{\link{getCmMetaEstimation}()},
\code{\link{getSccsDiagnosticsData}()},
\code{\link{getSccsEstimation}()}
}
\concept{{Estimation}}
