% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimationQueries.R
\name{getSccsTable}
\alias{getSccsTable}
\title{Extract the SCCS table specified}
\usage{
getSccsTable(
  connectionHandler,
  schema,
  table = c("attrition", "time_trend", "event_dep_observation", "age_spanning",
    "calendar_time_spanning", "spline")[1],
  sccsTablePrefix = "sccs_",
  cgTablePrefix = "cg_",
  databaseTable = "database_meta_data",
  indicationIds = NULL,
  outcomeIds = NULL,
  analysisIds = NULL,
  databaseIds = NULL,
  exposureOutcomeIds = NULL,
  covariateIds = NULL
)
}
\arguments{
\item{connectionHandler}{A connection handler that connects to the database and extracts sql queries.  Create a connection handler via `ResultModelManager::ConnectionHandler$new()`.}

\item{schema}{The result database schema (e.g., 'main' for sqlite)}

\item{table}{The result table to extract}

\item{sccsTablePrefix}{The prefix used for the cohort generator results tables}

\item{cgTablePrefix}{The prefix used for the cohort generator results tables}

\item{databaseTable}{The name of the table with the database details (default 'database_meta_data')}

\item{indicationIds}{The indications that the target was nested to}

\item{outcomeIds}{A vector of integers corresponding to the outcome cohort IDs}

\item{analysisIds}{the analysis IDs to restrict to}

\item{databaseIds}{the database IDs to restrict to}

\item{exposureOutcomeIds}{the exposureOutcomeIds to restrict to}

\item{covariateIds}{the covariateIds to restrict to}
}
\value{
Returns a data.frame with the cohort method requested table
}
\description{
This function extracts the specific cohort method table.
}
\details{
Specify the connectionHandler, the schema and optionally the target/outcome/analysis/database IDs
}
\examples{
conDet <- getExampleConnectionDetails()

connectionHandler <- ResultModelManager::ConnectionHandler$new(conDet)

sccsTable <- getSccsTable(
  connectionHandler = connectionHandler, 
  schema = 'main',
  table = 'attrition'
)

}
\seealso{
Other Estimation:
\code{\link[=.getCmVersion]{.getCmVersion()}},
\code{\link[=.getSccsVersion]{.getSccsVersion()}},
\code{\link[=getCMEstimation]{getCMEstimation()}},
\code{\link[=getCmDiagnosticsData]{getCmDiagnosticsData()}},
\code{\link[=getCmMetaEstimation]{getCmMetaEstimation()}},
\code{\link[=getCmNegativeControlEstimates]{getCmNegativeControlEstimates()}},
\code{\link[=getCmOutcomes]{getCmOutcomes()}},
\code{\link[=getCmPropensityModel]{getCmPropensityModel()}},
\code{\link[=getCmTable]{getCmTable()}},
\code{\link[=getCmTargets]{getCmTargets()}},
\code{\link[=getSccsDiagnosticsData]{getSccsDiagnosticsData()}},
\code{\link[=getSccsEstimation]{getSccsEstimation()}},
\code{\link[=getSccsMetaEstimation]{getSccsMetaEstimation()}},
\code{\link[=getSccsModel]{getSccsModel()}},
\code{\link[=getSccsNegativeControlEstimates]{getSccsNegativeControlEstimates()}},
\code{\link[=getSccsOutcomes]{getSccsOutcomes()}},
\code{\link[=getSccsTargets]{getSccsTargets()}},
\code{\link[=getSccsTimeToEvent]{getSccsTimeToEvent()}},
\code{\link[=plotCmEstimates]{plotCmEstimates()}},
\code{\link[=plotSccsEstimates]{plotSccsEstimates()}}
}
\concept{Estimation}
