% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortQueries.R
\name{getSubsetText}
\alias{getSubsetText}
\title{Function that converts a subsetDefinitionJson into text description}
\usage{
getSubsetText(subsetDefinitionJson, cohortDefinitions)
}
\arguments{
\item{subsetDefinitionJson}{The subset logic json}

\item{cohortDefinitions}{A data.frame with the columns cohortDefinitionId, cohortName and optionally friendlyName that will
be used to know the friendly cohort name for any subsetting that nests in other cohorts}
}
\value{
A text string describing the subsetting
}
\description{
Function that converts a subsetDefinitionJson into text description
}
\details{
The function takes a subsetDefinitionJson and converts it into friendly text describing the 
subset logic
}
\seealso{
Other Cohorts: 
\code{\link{getCohortAttrition}()},
\code{\link{getCohortCounts}()},
\code{\link{getCohortDefinitions}()},
\code{\link{getCohortInclusionRules}()},
\code{\link{getCohortInclusionStats}()},
\code{\link{getCohortInclusionSummary}()},
\code{\link{getCohortMeta}()},
\code{\link{getCohortSubsetAttrition}()},
\code{\link{getCohortSubsetDefinitions}()},
\code{\link{processCohortDefinitionsForQuarto}()},
\code{\link{processCohorts}()},
\code{\link{restrictCohortDefinitionsForQuarto}()}
}
\concept{Cohorts}
