% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharacterzationQueries.R
\name{getTargetBinaryFeatures}
\alias{getTargetBinaryFeatures}
\title{Extract aggregate statistics of binary feature analysis IDs of interest for targets}
\usage{
getTargetBinaryFeatures(
  connectionHandler,
  schema,
  cTablePrefix = "c_",
  cgTablePrefix = "cg_",
  databaseTable = "database_meta_data",
  targetIds = NULL,
  outcomeIds = NULL,
  analysisIds = c(3)
)
}
\arguments{
\item{connectionHandler}{A connection handler that connects to the database and extracts sql queries.  Create a connection handler via `ResultModelManager::ConnectionHandler$new()`.}

\item{schema}{The result database schema (e.g., 'main' for sqlite)}

\item{cTablePrefix}{The prefix used for the characterization results tables}

\item{cgTablePrefix}{The prefix used for the cohort generator results tables}

\item{databaseTable}{The name of the table with the database details (default 'database_meta_data')}

\item{targetIds}{A vector of integers corresponding to the target cohort IDs}

\item{outcomeIds}{A vector of integers corresponding to the outcome cohort IDs}

\item{analysisIds}{The feature extraction analysis ID of interest (e.g., 201 is condition)}
}
\value{
Returns a data.frame with the columns:
\itemize{
 \item{databaseName} \description{the name of the database}
 \item{targetName} \description{the target cohort name}
 \item{targetId} \description{the target cohort unique identifier}
 \item{outcomeName} \description{the outcome name}
 \item{outcomeId} \description{the outcome unique identifier}
 \item{minPriorObservation} \description{the minimum required observation days prior to index for an entry}
 \item{outcomeWashoutDays} \description{patients with the outcome occurring within this number of days prior to index are excluded (NA means no exclusion)}
 \item{covariateName} \description{the name of the feature}
 \item{sumValue} \description{the number of cases who have the feature value of 1}
}
}
\description{
This function extracts the feature extraction results for targets corresponding to specified target and outcome cohorts.
}
\details{
Specify the connectionHandler, the schema and the target/outcome cohort IDs
}
\seealso{
Other {Characterization}: 
\code{\link{getCaseBinaryFeatures}()},
\code{\link{getCaseCounts}()},
\code{\link{getCharacterizationDemographics}()},
\code{\link{getCohortDefinitions}()},
\code{\link{getDechallengeRechallenge}()},
\code{\link{getIncidenceRates}()},
\code{\link{getRiskFactors}()},
\code{\link{getTargetCounts}()},
\code{\link{getTimeToEvent}()},
\code{\link{plotAgeDistributions}()},
\code{\link{plotSexDistributions}()}
}
\concept{{Characterization}}
