% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharacterzationQueries.R
\name{getTargetsUsedInCharacterization}
\alias{getTargetsUsedInCharacterization}
\title{A function to extarct the targets found in characterization}
\usage{
getTargetsUsedInCharacterization(
  connectionHandler,
  schema,
  cTablePrefix = "c_",
  cgTablePrefix = "cg_",
  printTimes = FALSE,
  useTte = TRUE,
  useDcrc = TRUE,
  useRf = TRUE,
  useTb = TRUE,
  useCs = TRUE
)
}
\arguments{
\item{connectionHandler}{A connection handler that connects to the database and extracts sql queries.  Create a connection handler via `ResultModelManager::ConnectionHandler$new()`.}

\item{schema}{The result database schema (e.g., 'main' for sqlite)}

\item{cTablePrefix}{The prefix used for the characterization results tables}

\item{cgTablePrefix}{The prefix used for the cohort generator results tables}

\item{printTimes}{Print the time it takes to run each query}

\item{useTte}{whether to determine what cohorts are used in time to event}

\item{useDcrc}{whether to determine what cohorts are used in dechal-rechal}

\item{useRf}{whether to determine what cohorts are used in risk factor}

\item{useTb}{whether to determine what cohorts are used in target baseline}

\item{useCs}{whether to determine what cohorts are used in case-series}
}
\value{
A data.frame with the characterization target cohort ids, names and which characterization analyses the cohorts are used in.
}
\description{
A function to extarct the targets found in characterization
}
\details{
Specify the connectionHandler, the schema and the prefixes
}
\examples{
conDet <- getExampleConnectionDetails()

connectionHandler <- ResultModelManager::ConnectionHandler$new(conDet)

cohorts <- getTargetsUsedInCharacterization(
  connectionHandler = connectionHandler, 
  schema = 'main'
)

}
\seealso{
Other Characterization:
\code{\link[=characterizationCompareBinary]{characterizationCompareBinary()}},
\code{\link[=characterizationCompareContinuous]{characterizationCompareContinuous()}},
\code{\link[=getBinaryCaseSeries]{getBinaryCaseSeries()}},
\code{\link[=getBinaryRiskFactors]{getBinaryRiskFactors()}},
\code{\link[=getBinaryTargetBaseline]{getBinaryTargetBaseline()}},
\code{\link[=getCaseCounts]{getCaseCounts()}},
\code{\link[=getCharacterizationCaseSettings]{getCharacterizationCaseSettings()}},
\code{\link[=getCharacterizationDemographics]{getCharacterizationDemographics()}},
\code{\link[=getCharacterizationTargetSettings]{getCharacterizationTargetSettings()}},
\code{\link[=getContinuousCaseSeries]{getContinuousCaseSeries()}},
\code{\link[=getContinuousRiskFactors]{getContinuousRiskFactors()}},
\code{\link[=getContinuousTargetBaseline]{getContinuousTargetBaseline()}},
\code{\link[=getDechallengeRechallenge]{getDechallengeRechallenge()}},
\code{\link[=getDechallengeRechallengeFails]{getDechallengeRechallengeFails()}},
\code{\link[=getIncidenceRates]{getIncidenceRates()}},
\code{\link[=getIncidenceTargetSettings]{getIncidenceTargetSettings()}},
\code{\link[=getNonCaseCounts]{getNonCaseCounts()}},
\code{\link[=getOutcomesUsedInCharacterization]{getOutcomesUsedInCharacterization()}},
\code{\link[=getOutcomesUsedInIncidence]{getOutcomesUsedInIncidence()}},
\code{\link[=getTargetCounts]{getTargetCounts()}},
\code{\link[=getTargetsUsedInIncidence]{getTargetsUsedInIncidence()}},
\code{\link[=getTimeToEvent]{getTimeToEvent()}},
\code{\link[=plotAgeDistributions]{plotAgeDistributions()}},
\code{\link[=plotSexDistributions]{plotSexDistributions()}},
\code{\link[=viewIncidenceRate]{viewIncidenceRate()}}
}
\concept{Characterization}
