% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharacterzationQueries.R
\name{getTimeToEvent}
\alias{getTimeToEvent}
\title{Extract the time to event result}
\usage{
getTimeToEvent(
  connectionHandler,
  schema,
  cTablePrefix = "c_",
  cgTablePrefix = "cg_",
  databaseTable = "database_meta_data",
  targetIds = NULL,
  outcomeIds = NULL
)
}
\arguments{
\item{connectionHandler}{A connection handler that connects to the database and extracts sql queries.  Create a connection handler via `ResultModelManager::ConnectionHandler$new()`.}

\item{schema}{The result database schema (e.g., 'main' for sqlite)}

\item{cTablePrefix}{The prefix used for the characterization results tables}

\item{cgTablePrefix}{The prefix used for the cohort generator results tables}

\item{databaseTable}{The name of the table with the database details (default 'database_meta_data')}

\item{targetIds}{A vector of integers corresponding to the target cohort IDs}

\item{outcomeIds}{A vector of integers corresponding to the outcome cohort IDs}
}
\value{
Returns a data.frame with the columns:
\itemize{
 \item{databaseName} \description{the name of the database}
 \item{targetName} \description{the target cohort name}
 \item{targetId} \description{the target cohort unique identifier}
 \item{outcomeName} \description{the outcome name}
 \item{outcomeId} \description{the outcome unique identifier}
 \item{outcomeType} \description{Whether the outcome is the first or subsequent}
 \item{targetOutcomeType} \description{The interval that the outcome occurs}
 \item{timeToEvent} \description{The number of days from index}
 \item{numEvents} \description{The number of target cohort entries}
 \item{timeScale} \description{The correspondin time-scale}
 }
}
\description{
This function extracts all time to event results across databases for specified target and outcome cohorts.
}
\details{
Specify the connectionHandler, the schema and the target/outcome cohort IDs
}
\seealso{
Other {Characterization}: 
\code{\link{getCaseBinaryFeatures}()},
\code{\link{getCaseCounts}()},
\code{\link{getCharacterizationDemographics}()},
\code{\link{getCohortDefinitions}()},
\code{\link{getDechallengeRechallenge}()},
\code{\link{getIncidenceRates}()},
\code{\link{getRiskFactors}()},
\code{\link{getTargetBinaryFeatures}()},
\code{\link{getTargetCounts}()},
\code{\link{plotAgeDistributions}()},
\code{\link{plotSexDistributions}()}
}
\concept{{Characterization}}
