% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OhdsiReportGenerator.R
\docType{package}
\name{OhdsiReportGenerator}
\alias{OhdsiReportGenerator-package}
\alias{OhdsiReportGenerator}
\title{OhdsiReportGenerator}
\description{
A package for extracting analyses results and creating reports.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/OhdsiReportGenerator/}
  \item \url{https://github.com/OHDSI/OhdsiReportGenerator}
  \item Report bugs at \url{https://github.com/OHDSI/OhdsiReportGenerator/issues}
}

}
\author{
\strong{Maintainer}: Jenna Reps \email{jreps@its.jnj.com}

Authors:
\itemize{
  \item Anthony Sena \email{sena@ohdsi.org}
}

}
