% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharacterizationPlots.R
\name{plotAgeDistributions}
\alias{plotAgeDistributions}
\title{Plots the age distributions using the binary age groups}
\usage{
plotAgeDistributions(
  ageData,
  riskWindowStart = "1",
  riskWindowEnd = "365",
  startAnchor = "cohort start",
  endAnchor = "cohort start"
)
}
\arguments{
\item{ageData}{The age data extracted using 'getCharacterizationDemographics(type = 'age')'}

\item{riskWindowStart}{The time at risk window start}

\item{riskWindowEnd}{The time at risk window end}

\item{startAnchor}{The anchor for the time at risk start}

\item{endAnchor}{The anchor for the time at risk end}
}
\value{
Returns a ggplot with the distributions
}
\description{
Creates bar charts for the target and case age groups.
}
\details{
Input the data returned from 'getCharacterizationDemographics(type = 'age')' and the time-at-risk
}
\seealso{
Other Characterization: 
\code{\link{getCaseBinaryFeatures}()},
\code{\link{getCaseCounts}()},
\code{\link{getCharacterizationDemographics}()},
\code{\link{getDechallengeRechallenge}()},
\code{\link{getIncidenceRates}()},
\code{\link{getRiskFactors}()},
\code{\link{getTargetBinaryFeatures}()},
\code{\link{getTargetCounts}()},
\code{\link{getTimeToEvent}()},
\code{\link{plotSexDistributions}()}
}
\concept{Characterization}
