% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimationPlots.R
\name{plotCmEstimates}
\alias{plotCmEstimates}
\title{Plots the cohort method results for one analysis}
\usage{
plotCmEstimates(
  cmData,
  cmDiagnostics = NULL,
  cmMeta = NULL,
  cohortNames = NULL,
  includeCounts = TRUE,
  selectedAnalysisId = NULL
)
}
\arguments{
\item{cmData}{The cohort method data}

\item{cmDiagnostics}{(optional) The cohort method diagnostic data}

\item{cmMeta}{(optional) The cohort method evidence synthesis data}

\item{cohortNames}{A data.frame with columns cohortId and cohortName}

\item{includeCounts}{Whether to include the target/comp size and event counts}

\item{selectedAnalysisId}{The analysis ID of interest to plot}
}
\value{
Returns a ggplot with the estimates
}
\description{
Creates nice cohort method plots
}
\details{
Input the cohort method data
}
\examples{
conDet <- getExampleConnectionDetails()

connectionHandler <- ResultModelManager::ConnectionHandler$new(conDet)

cmEst <- getCMEstimation(
  connectionHandler = connectionHandler, 
  schema = 'main',
  targetIds = 1002,
  outcomeIds = 3
)
plotCmEstimates(
  cmData = cmEst, 
  cmMeta = NULL, 
  selectedAnalysisId = 1
)

}
\seealso{
Other Estimation: 
\code{\link{.getCmVersion}()},
\code{\link{.getSccsVersion}()},
\code{\link{getCMEstimation}()},
\code{\link{getCmDiagnosticsData}()},
\code{\link{getCmMetaEstimation}()},
\code{\link{getCmNegativeControlEstimates}()},
\code{\link{getCmOutcomes}()},
\code{\link{getCmPropensityModel}()},
\code{\link{getCmTable}()},
\code{\link{getCmTargets}()},
\code{\link{getSccsDiagnosticsData}()},
\code{\link{getSccsEstimation}()},
\code{\link{getSccsMetaEstimation}()},
\code{\link{getSccsModel}()},
\code{\link{getSccsNegativeControlEstimates}()},
\code{\link{getSccsOutcomes}()},
\code{\link{getSccsTable}()},
\code{\link{getSccsTargets}()},
\code{\link{getSccsTimeToEvent}()},
\code{\link{plotSccsEstimates}()}
}
\concept{Estimation}
