% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharacterizationPlots.R
\name{plotSexDistributions}
\alias{plotSexDistributions}
\title{Plots the sex distributions using the sex features}
\usage{
plotSexDistributions(sexData)
}
\arguments{
\item{sexData}{The sex data extracted using 'getCharacterizationDemographics(type = 'sex')'}
}
\value{
Returns a ggplot with the distributions
}
\description{
Creates bar charts for the target.
}
\details{
Input the data returned from 'getCharacterizationDemographics(type = 'sex')'
}
\examples{
conDet <- getExampleConnectionDetails()

connectionHandler <- ResultModelManager::ConnectionHandler$new(conDet)

sexData <- getCharacterizationDemographics(
  connectionHandler = connectionHandler, 
  schema = 'main',
  characterizationTargetId = 10, 
  outcomeId = 3, 
  type = 'sex'
)
plotSexDistributions(sexData = sexData)

}
\seealso{
Other Characterization:
\code{\link[=characterizationCompareBinary]{characterizationCompareBinary()}},
\code{\link[=characterizationCompareContinuous]{characterizationCompareContinuous()}},
\code{\link[=getAggregateBinaryRiskFactors]{getAggregateBinaryRiskFactors()}},
\code{\link[=getAggregateContinuousRiskFactors]{getAggregateContinuousRiskFactors()}},
\code{\link[=getBinaryCaseSeries]{getBinaryCaseSeries()}},
\code{\link[=getBinaryRiskFactors]{getBinaryRiskFactors()}},
\code{\link[=getBinaryTargetBaseline]{getBinaryTargetBaseline()}},
\code{\link[=getCaseCounts]{getCaseCounts()}},
\code{\link[=getCharacterizationCaseSettings]{getCharacterizationCaseSettings()}},
\code{\link[=getCharacterizationDemographics]{getCharacterizationDemographics()}},
\code{\link[=getCharacterizationTargetSettings]{getCharacterizationTargetSettings()}},
\code{\link[=getContinuousCaseSeries]{getContinuousCaseSeries()}},
\code{\link[=getContinuousRiskFactors]{getContinuousRiskFactors()}},
\code{\link[=getContinuousTargetBaseline]{getContinuousTargetBaseline()}},
\code{\link[=getDechallengeRechallenge]{getDechallengeRechallenge()}},
\code{\link[=getDechallengeRechallengeFails]{getDechallengeRechallengeFails()}},
\code{\link[=getIncidenceRates]{getIncidenceRates()}},
\code{\link[=getIncidenceTargetSettings]{getIncidenceTargetSettings()}},
\code{\link[=getNonCaseCounts]{getNonCaseCounts()}},
\code{\link[=getOutcomesUsedInCharacterization]{getOutcomesUsedInCharacterization()}},
\code{\link[=getOutcomesUsedInIncidence]{getOutcomesUsedInIncidence()}},
\code{\link[=getTargetCounts]{getTargetCounts()}},
\code{\link[=getTargetsUsedInCharacterization]{getTargetsUsedInCharacterization()}},
\code{\link[=getTargetsUsedInIncidence]{getTargetsUsedInIncidence()}},
\code{\link[=getTimeToEvent]{getTimeToEvent()}},
\code{\link[=plotAgeDistributions]{plotAgeDistributions()}},
\code{\link[=viewIncidenceRate]{viewIncidenceRate()}}
}
\concept{Characterization}
