% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharacterizationPlots.R
\name{plotSexDistributions}
\alias{plotSexDistributions}
\title{Plots the sex distributions using the sex features}
\usage{
plotSexDistributions(
  sexData,
  riskWindowStart = "1",
  riskWindowEnd = "365",
  startAnchor = "cohort start",
  endAnchor = "cohort start"
)
}
\arguments{
\item{riskWindowStart}{The time at risk window start}

\item{riskWindowEnd}{The time at risk window end}

\item{startAnchor}{The anchor for the time at risk start}

\item{endAnchor}{The anchor for the time at risk end}

\item{ageData}{The sex data extracted using 'getCharacterizationDemographics(type = 'sex')'}
}
\value{
Returns a ggplot with the distributions
}
\description{
Creates bar charts for the target and case sex.
}
\details{
Input the data returned from 'getCharacterizationDemographics(type = 'sex')' and the time-at-risk
}
\seealso{
Other {Characterization}: 
\code{\link{getCaseBinaryFeatures}()},
\code{\link{getCaseCounts}()},
\code{\link{getCharacterizationDemographics}()},
\code{\link{getCohortDefinitions}()},
\code{\link{getDechallengeRechallenge}()},
\code{\link{getIncidenceRates}()},
\code{\link{getRiskFactors}()},
\code{\link{getTargetBinaryFeatures}()},
\code{\link{getTargetCounts}()},
\code{\link{getTimeToEvent}()},
\code{\link{plotAgeDistributions}()}
}
\concept{{Characterization}}
