% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortQueries.R
\name{processCohortDefinitionsForQuarto}
\alias{processCohortDefinitionsForQuarto}
\title{Function processes the cohortDefinitions object ready for use in the main quarto report.}
\usage{
processCohortDefinitionsForQuarto(
  cohortDefinitions,
  friendlyCohortIds,
  friendlyCohortNames,
  restrictTargetToIndications,
  indicationIds
)
}
\arguments{
\item{cohortDefinitions}{The output of `getCohortDefinitions()`}

\item{friendlyCohortIds}{a vector of parent cohort ids that you want to rename}

\item{friendlyCohortNames}{a vector of new names for the friendlyCohortIds}

\item{restrictTargetToIndications}{whether to restrict the results to cohorts nested in certain indications}

\item{indicationIds}{The indication ids of interest for restrictTargetToIndications}
}
\value{
A cohortDefinitions object with extra columns: friendlyName,
}
\description{
Function processes the cohortDefinitions object ready for use in the main quarto report.
}
\details{
Function processes the cohortDefinitions object by adding friendly names for 
specified parent cohorts, determines which cohorts are nested in the specified
indication cohort ids and ...
}
\seealso{
Other Cohorts:
\code{\link[=getCohortAttrition]{getCohortAttrition()}},
\code{\link[=getCohortCounts]{getCohortCounts()}},
\code{\link[=getCohortDefinitions]{getCohortDefinitions()}},
\code{\link[=getCohortInclusionRules]{getCohortInclusionRules()}},
\code{\link[=getCohortInclusionStats]{getCohortInclusionStats()}},
\code{\link[=getCohortInclusionSummary]{getCohortInclusionSummary()}},
\code{\link[=getCohortMeta]{getCohortMeta()}},
\code{\link[=getCohortSubsetAttrition]{getCohortSubsetAttrition()}},
\code{\link[=getCohortSubsetDefinitions]{getCohortSubsetDefinitions()}},
\code{\link[=getSubsetText]{getSubsetText()}},
\code{\link[=processCohorts]{processCohorts()}},
\code{\link[=restrictCohortDefinitionsForQuarto]{restrictCohortDefinitionsForQuarto()}}
}
\concept{Cohorts}
