% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{removeSpaces}
\alias{removeSpaces}
\title{removeSpaces}
\usage{
removeSpaces(x)
}
\arguments{
\item{x}{A string}
}
\value{
A string without spaces
}
\description{
Removes spaces and replaces with under scroll
}
\details{
Removes spaces and replaces with under scroll
}
\examples{
removeSpaces(' made up.   string')

}
\seealso{
Other helper: 
\code{\link{getExampleConnectionDetails}()},
\code{\link{kableDark}()},
\code{\link{printReactable}()}
}
\concept{helper}
