% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{removeSpaces}
\alias{removeSpaces}
\title{removeSpaces}
\usage{
removeSpaces(x)
}
\arguments{
\item{x}{A string}
}
\value{
A string without spaces
}
\description{
Removes spaces and replaces with under scroll
}
\details{
Removes spaces and replaces with under scroll
}
\examples{
removeSpaces(' made up.   string')

}
\seealso{
Other helper:
\code{\link[=addTarColumn]{addTarColumn()}},
\code{\link[=formatBinaryCovariateName]{formatBinaryCovariateName()}},
\code{\link[=formatTargetPopulation]{formatTargetPopulation()}},
\code{\link[=getExampleConnectionDetails]{getExampleConnectionDetails()}},
\code{\link[=getOutcomeTable]{getOutcomeTable()}},
\code{\link[=getTarText]{getTarText()}},
\code{\link[=getTargetTable]{getTargetTable()}},
\code{\link[=kableDark]{kableDark()}}
}
\concept{helper}
